+++ leo/core/leoserver.py
@@ -1,1592 +1,831 @@
-#! python3
 #@+leo-ver=5-thin
-#@+node:ekr.20210611084045.1: * @file leoserver.py
-#@@first
-"""
-leoserver.py: same as leointeg/leobridgeserver.py except for sys.path manipulations.
-"""
+#@+node:ekr.20210611081444.1: * @file leoserver.py
 #@@language python
 #@@tabwidth -4
+"""
+Leo's internet server. (The beauty queen). Does not work yet with leoInteg.
+
+Written by Félix Malboeuf and Edward K. Ream.
+"""
+# pylint: disable=import-self,raise-missing-from
 #@+<< imports >>
-#@+node:ekr.20210611084045.2: ** << imports >>
-# pylint: disable=wrong-import-position
+#@+node:ekr.20210202110128.2: ** << imports >>
 import asyncio
 import getopt
-# import inspect
+import inspect
 import json
 import os
-import os.path
 import sys
 import time
 import unittest
-import traceback
 # Third-party.
 import websockets
-# #1970: Adjust sys.path so that the following imports will always work.
-_leo_editor_path = os.path.abspath(os.path.join(__file__, '..', '..', '..'))
-if _leo_editor_path not in sys.path:
-    sys.path.append(_leo_editor_path)
 # Leo
-import leo.core.leoBridge as leoBridge
-import leo.core.leoNodes as leoNodes
+from leo.core.leoNodes import Position
 from leo.core.leoGui import StringFindTabManager
+from leo.core import leoserver
 #@-<< imports >>
-#@+<< constants >>
-#@+node:ekr.20210611084045.3: ** << constants >>
-# server defaults
+g = None  # The bridge's leoGlobals module. Unit tests use self.g.
+# For unit tests.
+g_leoserver = None
+g_server = None
+# Server defaults...
 wsHost = "localhost"
 wsPort = 32125
 
-# To help in printout
-commonActions = ["getChildren", "getBody", "getBodyLength"]
-
-# Special string signals server startup success
-SERVER_STARTED_TOKEN = "LeoBridge started"
-
-
-#@-<< constants >>
 #@+others
-#@+node:ekr.20210611084045.4: ** class IdleTimeManager
-class IdleTimeManager:
-    """
-    A singleton class to manage idle-time handling. This class handles all
-    details of running code at idle time, including running 'idle' hooks.
+#@+node:ekr.20210204054519.1: ** Exception classes
+class InternalServerError(Exception):  # pragma: no cover
+    """The server violated its own coding conventions."""
+    pass
 
-    Any code can call g.app.idleTimeManager.add_callback(callback) to cause
-    the callback to be called at idle time forever.
-    """
-    # TODO : REVISE/REPLACE WITH OWN SYSTEM
+class ServerError(Exception):  # pragma: no cover
+    """The server received an erroneous package."""
+    pass
 
-    def __init__(self, g):
-        """Ctor for IdleTimeManager class."""
-        self.g = g
-        self.callback_list = []
-        self.timer = None
-        self.on_idle_count = 0
-
+class TerminateServer(Exception):  # pragma: no cover
+    """Ask the server to terminate."""
+    pass
+#@+node:ekr.20210202110128.29: ** class LeoServer
+class LeoServer:
+    """Leo Server Controller"""
     #@+others
-    #@+node:ekr.20210611084045.5: *3* itm.add_callback
-    def add_callback(self, callback):
-        """Add a callback to be called at every idle time."""
-        self.callback_list.append(callback)
+    #@+node:ekr.20210202110128.30: *3* server.__init__ (load bridge)
+    def __init__(self, testing=False):
 
-    #@+node:ekr.20210611084045.6: *3* itm.on_idle
-    def on_idle(self, timer):
-        """IdleTimeManager: Run all idle-time callbacks."""
-        if not self.g.app:
-            return
-        if self.g.app.killed:
-            return
-        if not self.g.app.pluginsController:
-            self.g.trace('No g.app.pluginsController', self.g.callers())
-            timer.stop()
-            return  # For debugger.
-        self.on_idle_count += 1
-        # Handle the registered callbacks.
-        # print("list length : ", len(self.callback_list))
-        for callback in self.callback_list:
-            try:
-                callback()
-            except Exception:
-                self.g.es_exception()
-                self.g.es_print(f"removing callback: {callback}")
-                self.callback_list.remove(callback)
-        # Handle idle-time hooks.
-        self.g.app.pluginsController.on_idle()
-
-    #@+node:ekr.20210611084045.7: *3* itm.start
-    def start(self):
-        """Start the idle-time timer."""
-        self.timer = self.g.IdleTime(
-            self.on_idle,
-            delay=500,  # Milliseconds
-            tag='IdleTimeManager.on_idle')
-        if self.timer:
-            self.timer.start()
-
-    #@-others
-
-#@+node:ekr.20210611084045.8: ** class ExternalFilesController
-class ExternalFilesController:
-    '''EFC Modified from Leo's sources'''
-    # pylint: disable=no-else-return
-
-    #@+others
-    #@+node:ekr.20210611084045.9: *3* efc.ctor
-    def __init__(self, integController):
-        '''Ctor for ExternalFiles class.'''
-        self.on_idle_count = 0
-        self.integController = integController
-        self.checksum_d = {}
-        # Keys are full paths, values are file checksums.
-        self.enabled_d = {}
-        # For efc.on_idle.
-        # Keys are commanders.
-        # Values are cached @bool check-for-changed-external-file settings.
-        self.has_changed_d = {}
-        # Keys are commanders. Values are boolean.
-        # Used only to limit traces.
-        self.unchecked_commanders = []
-        # Copy of g.app.commanders()
-        self.unchecked_files = []
-        # Copy of self file. Only one files is checked at idle time.
-        self._time_d = {}
-        # Keys are full paths, values are modification times.
-        # DO NOT alter directly, use set_time(path) and
-        # get_time(path), see set_time() for notes.
-        self.yesno_all_time = 0  # previous yes/no to all answer, time of answer
-        self.yesno_all_answer = None  # answer, 'yes-all', or 'no-all'
-
-        # if yesAll/noAll forced, then just show info message after idle_check_commander
-        self.infoMessage = None
-        # False or "detected", "refreshed" or "ignored"
-
-        self.integController.g.app.idleTimeManager.add_callback(self.on_idle)
-
-        self.waitingForAnswer = False
-        self.lastPNode = None  # last p node that was asked for if not set to "AllYes\AllNo"
-        self.lastCommander = None
-
-    #@+node:ekr.20210611084045.10: *3* efc.on_idle
-    def on_idle(self):
-        '''
-        Check for changed open-with files and all external files in commanders
-        for which @bool check_for_changed_external_file is True.
-        '''
-        # Fix for flushing the terminal console to traverse
-        # python through node.js when using start server in leoInteg
-        sys.stdout.flush()
-
-        if not self.integController.g.app or self.integController.g.app.killed:
-            return
-        if self.waitingForAnswer:
-            return
-
-        self.on_idle_count += 1
-
-        if self.unchecked_commanders:
-            # Check the next commander for which
-            # @bool check_for_changed_external_file is True.
-            c = self.unchecked_commanders.pop()
-            self.lastCommander = c
-            self.idle_check_commander(c)
-        else:
-            # Add all commanders for which
-            # @bool check_for_changed_external_file is True.
-            self.unchecked_commanders = [
-                z for z in self.integController.g.app.commanders() if self.is_enabled(z)
-            ]
-
-    #@+node:ekr.20210611084045.11: *3* efc.idle_check_commander
-    def idle_check_commander(self, c):
-        '''
-        Check all external files corresponding to @<file> nodes in c for
-        changes.
-        '''
-        # #1100: always scan the entire file for @<file> nodes.
-        # #1134: Nested @<file> nodes are no longer valid, but this will do no harm.
-
-        self.infoMessage = None  # reset infoMessage
-        # False or "detected", "refreshed" or "ignored"
-
-        for p in c.all_unique_positions():
-            if self.waitingForAnswer:
-                break
-            if p.isAnyAtFileNode():
-                self.idle_check_at_file_node(c, p)
-
-        # if yesAll/noAll forced, then just show info message
-        if self.infoMessage:
-            w_package = {"async": "info", "message": self.infoMessage}
-            self.integController.sendAsyncOutput(w_package)
-
-    #@+node:ekr.20210611084045.12: *3* efc.idle_check_at_file_node
-    def idle_check_at_file_node(self, c, p):
-        '''Check the @<file> node at p for external changes.'''
-        trace = False
-        # Matt, set this to True, but only for the file that interests you.\
-        # trace = p.h == '@file unregister-leo.leox'
-        path = self.integController.g.fullPath(c, p)
-        has_changed = self.has_changed(path)
-        if trace:
-            self.integController.g.trace('changed', has_changed, p.h)
-        if has_changed:
-            self.lastPNode = p  # can be set here because its the same process for ask/warn
-            if p.isAtAsisFileNode() or p.isAtNoSentFileNode():
-                # Fix #1081: issue a warning.
-                self.warn(c, path, p=p)
-            elif self.ask(c, path, p=p):
-                self.lastCommander.selectPosition(self.lastPNode)
-                c.refreshFromDisk()
-
-            # Always update the path & time to prevent future warnings.
-            self.set_time(path)
-            self.checksum_d[path] = self.checksum(path)
-
-    #@+node:ekr.20210611084045.13: *3* efc.integResult
-    def integResult(self, p_result):
-        '''Received result from client'''
-        # Got the result to an asked question/warning from the client
-        if not self.waitingForAnswer:
-            print("ERROR: Received Result but no Asked Dialog", flush=True)
-            return
-        # check if p_resultwas from a warn (ok) or an ask ('yes','yes-all','no','no-all')
-        # act accordingly
-
-        path = self.integController.g.fullPath(
-            self.lastCommander, self.lastPNode)
-
-        # 1- if ok, unblock 'warn'
-        # 2- if no, unblock 'ask'
-        # ------------------------------------------ Nothing special to do
-
-        # 3- if noAll, set noAll, and unblock 'ask'
-        if p_result and "-all" in p_result.lower():
-            self.yesno_all_time = time.time()
-            self.yesno_all_answer = p_result.lower()
-        # ------------------------------------------ Also covers setting yesAll in #5
-
-        # 4- if yes, REFRESH self.lastPNode, and unblock 'ask'
-        # 5- if yesAll,REFRESH self.lastPNode, set yesAll, and unblock 'ask'
-        if bool(p_result and 'yes' in p_result.lower()):
-            self.lastCommander.selectPosition(self.lastPNode)
-            self.lastCommander.refreshFromDisk()
-
-        # Always update the path & time to prevent future warnings for this PNode.
-        self.set_time(path)
-        self.checksum_d[path] = self.checksum(path)
-
-        self.waitingForAnswer = False  # unblock
-        # unblock: run the loop as if timer had hit
-        self.idle_check_commander(self.lastCommander)
-
-    #@+node:ekr.20210611084045.14: *3* efc.utilities
-    #@+node:ekr.20210611084045.15: *4* efc.ask
-    def ask(self, c, path, p=None):
-        '''
-        Ask user whether to overwrite an @<file> tree.
-        Return True if the user agrees.
-        '''
-        # check with leoInteg's config first
-        if self.integController.leoIntegConfig:
-            w_check_config = self.integController.leoIntegConfig["defaultReloadIgnore"].lower(
-            )
-            if not bool('none' in w_check_config):
-                if bool('yes' in w_check_config):
-                    self.infoMessage = "refreshed"
-                    return True
-                else:
-                    self.infoMessage = "ignored"
-                    return False
-        # let original function resolve
-
-        if self.yesno_all_time + 3 >= time.time() and self.yesno_all_answer:
-            self.yesno_all_time = time.time()  # Still reloading?  Extend time.
-            # if yesAll/noAll forced, then just show info message
-            w_yesno_all_bool = bool('yes' in self.yesno_all_answer.lower())
-
-            return w_yesno_all_bool
-        if not p:
-            where = 'the outline node'
-        else:
-            where = p.h
-
-        _is_leo = path.endswith(('.leo', '.db'))
-
-        if _is_leo:
-            s = '\n'.join([
-                f'{self.integController.g.splitLongFileName(path)} has changed outside Leo.',
-                'Overwrite it?'
-            ])
-        else:
-            s = '\n'.join([
-                f'{self.integController.g.splitLongFileName(path)} has changed outside Leo.',
-                f"Reload {where} in Leo?",
-            ])
-
-        w_package = {"async": "ask", "ask": 'Overwrite the version in Leo?',
-                     "message": s, "yes_all": not _is_leo, "no_all": not _is_leo}
-
-        self.integController.sendAsyncOutput(w_package)
-        self.waitingForAnswer = True
-        return False
-        # result = self.integController.g.app.gui.runAskYesNoDialog(c, 'Overwrite the version in Leo?', s,
-        # yes_all=not _is_leo, no_all=not _is_leo)
-
-        # if result and "-all" in result.lower():
-        # self.yesno_all_time = time.time()
-        # self.yesno_all_answer = result.lower()
-
-        # return bool(result and 'yes' in result.lower())
-
-    #@+node:ekr.20210611084045.16: *4* efc.checksum
-    def checksum(self, path):
-        '''Return the checksum of the file at the given path.'''
-        import hashlib
-        return hashlib.md5(open(path, 'rb').read()).hexdigest()
-
-    #@+node:ekr.20210611084045.17: *4* efc.get_mtime
-    def get_mtime(self, path):
-        '''Return the modification time for the path.'''
-        return self.integController.g.os_path_getmtime(self.integController.g.os_path_realpath(path))
-
-    #@+node:ekr.20210611084045.18: *4* efc.get_time
-    def get_time(self, path):
-        '''
-        return timestamp for path
-
-        see set_time() for notes
-        '''
-        return self._time_d.get(self.integController.g.os_path_realpath(path))
-
-    #@+node:ekr.20210611084045.19: *4* efc.has_changed
-    def has_changed(self, path):
-        '''Return True if p's external file has changed outside of Leo.'''
-        if not path:
-            return False
-        if not self.integController.g.os_path_exists(path):
-            return False
-        if self.integController.g.os_path_isdir(path):
-            return False
+        import leo.core.leoApp as leoApp
+        import leo.core.leoBridge as leoBridge
+        import leo.core.leoExternalFiles as leoExternalFiles
+        global g
+        t1 = time.process_time()
         #
-        # First, check the modification times.
-        old_time = self.get_time(path)
-        new_time = self.get_mtime(path)
-        if not old_time:
-            # Initialize.
-            self.set_time(path, new_time)
-            self.checksum_d[path] = self.checksum(path)
-            return False
-        if old_time == new_time:
-            return False
+        # Init ivars first.
+        self.c = None  # Currently Selected Commander.
+        self.dummy_c = None  # Set below, after we set g.
+        self.action = None
+        self.bad_commands_list = []  # Set below.
+        self.config = None
+        self.current_id = 0  # Id of action being processed.
+        self.log_flag = False  # set by "log" key
         #
-        # Check the checksums *only* if the mod times don't match.
-        old_sum = self.checksum_d.get(path)
-        new_sum = self.checksum(path)
-        if new_sum == old_sum:
-            # The modtime changed, but it's contents didn't.
-            # Update the time, so we don't keep checking the checksums.
-            # Return False so we don't prompt the user for an update.
-            self.set_time(path, new_time)
-            return False
-        # The file has really changed.
-        assert old_time, path
-        # #208: external change overwrite protection only works once.
-        # If the Leo version is changed (dirtied) again,
-        # overwrite will occur without warning.
-        # self.set_time(path, new_time)
-        # self.checksum_d[path] = new_sum
-        return True
-
-    #@+node:ekr.20210611084045.20: *4* efc.is_enabled
-    def is_enabled(self, c):
-        '''Return the cached @bool check_for_changed_external_file setting.'''
-        # check with leoInteg's config first
-        if self.integController.leoIntegConfig:
-            w_check_config = self.integController.leoIntegConfig["checkForChangeExternalFiles"].lower(
-            )
-            if bool('check' in w_check_config):
-                return True
-            if bool('ignore' in w_check_config):
-                return False
-        # let original function resolve
-        d = self.enabled_d
-        val = d.get(c)
-        if val is None:
-            val = c.config.getBool(
-                'check-for-changed-external-files', default=False)
-            d[c] = val
-        return val
-
-    #@+node:ekr.20210611084045.21: *4* efc.join
-    def join(self, s1, s2):
-        '''Return s1 + ' ' + s2'''
-        return f"{s1} {s2}"
-
-    #@+node:ekr.20210611084045.22: *4* efc.set_time
-    def set_time(self, path, new_time=None):
-        '''
-        Implements c.setTimeStamp.
-
-        Update the timestamp for path.
-
-        NOTE: file paths with symbolic links occur with and without those links
-        resolved depending on the code call path.  This inconsistency is
-        probably not Leo's fault but an underlying Python issue.
-        Hence the need to call realpath() here.
-        '''
-
-        # print("called set_time for " + str(path), flush=True)
-
-        t = new_time or self.get_mtime(path)
-        self._time_d[self.integController.g.os_path_realpath(path)] = t
-
-    #@+node:ekr.20210611084045.23: *4* efc.warn
-    def warn(self, c, path, p):
-        '''
-        Warn that an @asis or @nosent node has been changed externally.
-
-        There is *no way* to update the tree automatically.
-        '''
-        # check with leoInteg's config first
-        if self.integController.leoIntegConfig:
-            w_check_config = self.integController.leoIntegConfig["defaultReloadIgnore"].lower(
-            )
-
-            if w_check_config != "none":
-                # if not 'none' then do not warn, just infoMessage 'warn' at most
-                if not self.infoMessage:
-                    self.infoMessage = "warn"
-                return
-
-        # let original function resolve
-        if self.integController.g.unitTesting or c not in self.integController.g.app.commanders():
-            return
-        if not p:
-            self.integController.g.trace('NO P')
-            return
-
-        s = '\n'.join([
-            '%s has changed outside Leo.\n' % self.integController.g.splitLongFileName(
-                path),
-            'Leo can not update this file automatically.\n',
-            'This file was created from %s.\n' % p.h,
-            'Warning: refresh-from-disk will destroy all children.'
-        ])
-
-        w_package = {"async": "warn",
-                     "warn": 'External file changed', "message": s}
-
-        self.integController.sendAsyncOutput(w_package)
-        self.waitingForAnswer = True
-
-    #@+node:ekr.20210611084045.24: *3* other called methods
-    # Some methods are called in the usual (non-leoBridge without 'efc') save process.
-    # Those may be called by the 'save' function, like check_overwrite,
-    # or by any other functions from the instance of leo.core.leoBridge that's running.
-
-    #@+node:ekr.20210611084045.25: *4* open_with
-    def open_with(self, c, d):
-        return
-
-    #@+node:ekr.20210611084045.26: *4* check_overwrite
-    def check_overwrite(self, c, fn):
-        # print("check_overwrite!! ", flush=True)
-        return True
-
-    #@+node:ekr.20210611084045.27: *4* shut_down
-    def shut_down(self):
-        return
-
-    #@+node:ekr.20210611084045.28: *4* destroy_frame
-    def destroy_frame(self, f):
-        return
-
-    #@-others
-
-#@+node:ekr.20210611084045.29: ** class IntegTextWrapper
-class IntegTextWrapper:
-    """
-    A class that represents text as a Python string.
-    Modified from Leo's StringTextWrapper class source
-    """
-
-    #@+others
-    #@+node:ekr.20210611084045.30: *3* stw.ctor
-    def __init__(self, c, name, g):
-        """Ctor for the IntegTextWrapper class."""
-        self.c = c
-        self.name = name
-        self.g = g  # Should g be totally global across all leoIntegration classes?
-        self.ins = 0
-        self.sel = 0, 0
-        self.s = ''
-        self.yScroll = 0
-        self.supportsHighLevelInterface = True
-        self.widget = None  # This ivar must exist, and be None.
-
-    def __repr__(self):
-        return f"<IntegTextWrapper: {id(self)} {self.name}>"
-
-    def getName(self):
-        """IntegTextWrapper."""
-        return self.name  # Essential.
-
-    #@+node:ekr.20210611084045.31: *3* stw.Clipboard
-    def clipboard_clear(self):
-        self.g.app.gui.replaceClipboardWith('')
-
-    def clipboard_append(self, s):
-        s1 = self.g.app.gui.getTextFromClipboard()
-        self.g.app.gui.replaceClipboardWith(s1 + s)
-
-    #@+node:ekr.20210611084045.32: *3* stw.Do-nothings
-    def flashCharacter(self, i, bg='white', fg='red',
-                       flashes=3, delay=75): pass
-
-    def see(self, i): pass
-
-    def seeInsertPoint(self): pass
-
-    def setFocus(self): pass
-
-    def setStyleClass(self, name): pass
-
-    def tag_configure(self, colorName, **keys): pass
-
-    #@+node:ekr.20210611084045.33: *3* stw.Text
-    #@+node:ekr.20210611084045.34: *4* stw.appendText
-    def appendText(self, s):
-        """IntegTextWrapper appendText"""
-        self.s = self.s + self.g.toUnicode(s)
-        # defensive
-        self.ins = len(self.s)
-        self.sel = self.ins, self.ins
-
-    #@+node:ekr.20210611084045.35: *4* stw.delete
-    def delete(self, i, j=None):
-        """IntegTextWrapper delete"""
-        i = self.toPythonIndex(i)
-        if j is None:
-            j = i + 1
-        j = self.toPythonIndex(j)
-        # This allows subclasses to use this base class method.
-        if i > j:
-            i, j = j, i
-        s = self.getAllText()
-        self.setAllText(s[:i] + s[j:])
-        # Bug fix: 2011/11/13: Significant in external tests.
-        self.setSelectionRange(i, i, insert=i)
-
-    #@+node:ekr.20210611084045.36: *4* stw.deleteTextSelection
-    def deleteTextSelection(self):
-        """IntegTextWrapper."""
-        i, j = self.getSelectionRange()
-        self.delete(i, j)
-
-    #@+node:ekr.20210611084045.37: *4* stw.get
-    def get(self, i, j=None):
-        """IntegTextWrapper get"""
-        i = self.toPythonIndex(i)
-        if j is None:
-            j = i + 1
-        j = self.toPythonIndex(j)
-        s = self.s[i:j]
-        # print("WRAPPER GET with self.s[i:j]: " + s)
-        return self.g.toUnicode(s)
-
-    #@+node:ekr.20210611084045.38: *4* stw.getAllText
-    def getAllText(self):
-        """IntegTextWrapper getAllText"""
-        s = self.s
-        # print("WRAPPER getAllText  " + s)
-        return self.g.checkUnicode(s)
-
-    #@+node:ekr.20210611084045.39: *4* stw.getInsertPoint
-    def getInsertPoint(self):
-        """IntegTextWrapper getInsertPoint"""
-        i = self.ins
-        if i is None:
-            if self.virtualInsertPoint is None:
-                i = 0
-            else:
-                i = self.virtualInsertPoint
-        self.virtualInsertPoint = i
-        return i
-
-    #@+node:ekr.20210611084045.40: *4* stw.getSelectedText
-    def getSelectedText(self):
-        """IntegTextWrapper getSelectedText"""
-        i, j = self.sel
-        s = self.s[i:j]
-        # print("WRAPPER getSelectedText with self.s[i:j]: " + s)
-        return self.g.checkUnicode(s)
-
-    #@+node:ekr.20210611084045.41: *4* stw.getSelectionRange
-    def getSelectionRange(self, sort=True):
-        """Return the selected range of the widget."""
-        sel = self.sel
-        if len(sel) == 2 and sel[0] >= 0 and sel[1] >= 0:
-            i, j = sel
-            if sort and i > j:
-                sel = j, i  # Bug fix: 10/5/07
-            return sel
-        i = self.ins
-        return i, i
-
-    #@+node:ekr.20210611084045.42: *4* stw.getXScrollPosition
-    def getXScrollPosition(self):
-        return 0
-        # X axis ignored
-
-    #@+node:ekr.20210611084045.43: *4* stw.getYScrollPosition
-    def getYScrollPosition(self):
-        # print("wrapper get y scroll" + str(self.yScroll))
-        return self.yScroll
-
-    #@+node:ekr.20210611084045.44: *4* stw.hasSelection
-    def hasSelection(self):
-        """IntegTextWrapper hasSelection"""
-        i, j = self.getSelectionRange()
-        return i != j
-
-    #@+node:ekr.20210611084045.45: *4* stw.insert
-    def insert(self, i, s):
-        """IntegTextWrapper insert"""
-        i = self.toPythonIndex(i)
-        s1 = s
-        self.s = self.s[:i] + s1 + self.s[i:]
-        i += len(s1)
-        self.ins = i
-        self.sel = i, i
-
-    #@+node:ekr.20210611084045.46: *4* stw.selectAllText
-    def selectAllText(self, insert=None):
-        """IntegTextWrapper selectAllText"""
-        self.setSelectionRange(0, 'end', insert=insert)
-
-    #@+node:ekr.20210611084045.47: *4* stw.setAllText
-    def setAllText(self, s):
-        """IntegTextWrapper setAllText"""
-        # print("WRAPPER setAllText: " + s)
-        self.s = s
-        i = len(self.s)
-        self.ins = i
-        self.sel = i, i
-
-    #@+node:ekr.20210611084045.48: *4* stw.setInsertPoint
-    def setInsertPoint(self, pos, s=None):
-        """IntegTextWrapper setInsertPoint"""
-        self.virtualInsertPoint = i = self.toPythonIndex(pos)
-        self.ins = i
-        self.sel = i, i
-
-    #@+node:ekr.20210611084045.49: *4* stw.setXScrollPosition
-    def setXScrollPosition(self, i):
-        pass
-        # X axis ignored
-
-    #@+node:ekr.20210611084045.50: *4* stw.setYScrollPosition
-    def setYScrollPosition(self, i):
-        self.yScroll = i
-        # print("wrapper set y scroll" + str(self.yScroll))
-
-    #@+node:ekr.20210611084045.51: *4* stw.setSelectionRange
-    def setSelectionRange(self, i, j, insert=None):
-        """IntegTextWrapper setSelectionRange"""
-        i, j = self.toPythonIndex(i), self.toPythonIndex(j)
-        self.sel = i, j
-        self.ins = j if insert is None else self.toPythonIndex(insert)
-
-    #@+node:ekr.20210611084045.52: *4* stw.toPythonIndex
-    def toPythonIndex(self, index):
-        """IntegTextWrapper toPythonIndex"""
-        return self.g.toPythonIndex(self.s, index)
-
-    #@+node:ekr.20210611084045.53: *4* stw.toPythonIndexRowCol
-    def toPythonIndexRowCol(self, index):
-        """IntegTextWrapper toPythonIndexRowCol"""
-        s = self.getAllText()
-        i = self.toPythonIndex(index)
-        row, col = self.g.convertPythonIndexToRowCol(s, i)
-        return i, row, col
-
-    #@-others
-
-#@+node:ekr.20210611084045.54: ** class LeoBridgeIntegController
-class LeoBridgeIntegController:
-    '''Leo Bridge Controller'''
-    # pylint: disable=no-else-return
-
-    #@+others
-    #@+node:ekr.20210611084045.55: *3* __init__
-    def __init__(self):
-        self.gnx_to_vnode = []  # utility array - see leoflexx.py in leoPluginsRef.leo
+        # Start the bridge.
         self.bridge = leoBridge.controller(
             gui='nullGui',
-            loadPlugins=True,   # True: attempt to load plugins.
-            readSettings=True,  # True: read standard settings files.
-            silent=True,       # True: don't print signon messages.
-            # True: prints what would be sent to the log pane.
-            verbose=False,
+            loadPlugins=False,   # True: attempt to load plugins.
+            readSettings=False,  # True: read standard settings files.
+            silent=True,         # True: don't print signon messages.
+            verbose=False,       # True: prints messages that would be sent to the log pane.
         )
-        self.g = self.bridge.globals()
-
-        # * Trace outputs to pythons stdout, also prints the function call stack
-        # self.g.trace('test trace')
-
-        # * Intercept Log Pane output: Sends to client's log pane
-        self.g.es = self.es  # pointer - not a function call
-
-        # print(dir(self.g), flush=True)
-        self.currentActionId = 1  # Id of action being processed, STARTS AT 1 = Initial 'ready'
-
-        # * Currently Selected Commander (opened from leo.core.leoBridge or chosen via the g.app.windowList 2 list)
-        self.commander = None
-
-        self.leoIntegConfig = None
-        self.webSocket = None
-        self.loop = None
-
-        # * Replacement instances to Leo's codebase : getScript, IdleTime, idleTimeManager and externalFilesController
-        self.g.getScript = self._getScript
-        self.g.IdleTime = self._idleTime
-        self.g.app.idleTimeManager = IdleTimeManager(self.g)
-        # attach instance to g.app for calls to set_time, etc.
-        self.g.app.externalFilesController = ExternalFilesController(self)
-        # TODO : Maybe use those yes/no replacement right before actual usage instead of in init. (to allow re-use/switching)
-        # override for "revert to file" operation
-        self.g.app.gui.runAskYesNoDialog = self._returnYes
-        self.g.app.gui.show_find_success = self._show_find_success
-        self.headlineWidget = self.g.bunch(_name='tree')
-
-        # self.g.app.loadManager.createAllImporterData()  # Fixed in #1965 of leo-editor
-
-        # * setup leoBackground to get messages from leo
-        try:
-            self.g.app.idleTimeManager.start()  # To catch derived file changes
-        except Exception:
-            print('ERROR with idleTimeManager')
-
-    #@+node:ekr.20210611084045.56: *3* _asyncIdleLoop
-    async def _asyncIdleLoop(self, p_seconds, p_fn):
-        while True:
-            await asyncio.sleep(p_seconds)
-            p_fn(self)
-
-    #@+node:ekr.20210611084045.57: *3* _returnNo
-    def _returnNo(self, *arguments, **kwargs):
-        '''Used to override g.app.gui.ask[XXX] dialogs answers'''
-        return "no"
-
-    #@+node:ekr.20210611084045.58: *3* _returnYes
-    def _returnYes(self, *arguments, **kwargs):
-        '''Used to override g.app.gui.ask[XXX] dialogs answers'''
-        return "yes"
-
-    #@+node:ekr.20210611084045.59: *3* _getScript
-    def _getScript(self, c, p,
-                   useSelectedText=True,
-                   forcePythonSentinels=True,
-                   useSentinels=True,
-                   ):
-        """
-        Return the expansion of the selected text of node p.
-        Return the expansion of all of node p's body text if
-        p is not the current node or if there is no text selection.
-        """
-        w = c.frame.body.wrapper
-        if not p:
-            p = c.p
-        try:
-            if w and p == c.p and useSelectedText and w.hasSelection():
-                s = w.getSelectedText()
-            else:
-                s = p.b
-            # Remove extra leading whitespace so the user may execute indented code.
-            s = self.g.removeExtraLws(s, c.tab_width)
-            s = self.g.extractExecutableString(c, p, s)
-            script = self.g.composeScript(c, p, s,
-                                          forcePythonSentinels=forcePythonSentinels,
-                                          useSentinels=useSentinels)
-        except Exception:
-            self.g.es_print("unexpected exception in g.getScript")
-            self.g.es_exception()
-            script = ''
-        return script
-
-    #@+node:ekr.20210611084045.60: *3* _idleTime
-    def _idleTime(self, fn, delay, tag):
-        # TODO : REVISE/REPLACE WITH OWN SYSTEM
-        asyncio.get_event_loop().create_task(self._asyncIdleLoop(delay/1000, fn))
-
-    #@+node:ekr.20210611084045.61: *3* _getTotalOpened
-    def _getTotalOpened(self):
-        '''Get total of opened commander (who have closed == false)'''
-        w_total = 0
-        for w_commander in self.g.app.commanders():
-            if not w_commander.closed:
-                w_total = w_total + 1
-        return w_total
-
-        # return sum(1 for z in self.g.app.commanders() if not z.closed)
-    #@+node:ekr.20210611084045.62: *3* _getFirstOpenedCommander
-    def _getFirstOpenedCommander(self):
-        '''Get first opened commander, or False if there are none.'''
-        for w_commander in self.g.app.commanders():
-            if not w_commander.closed:
-                return w_commander
-        return False
-
-    #@+node:ekr.20210611084045.63: *3* _show_find_success
-    def _show_find_success(self, c, in_headline, insert, p):
-        '''Handle a successful find match.'''
-        if in_headline:
-            self.g.app.gui.set_focus(c, self.headlineWidget)
-        # no return
-
-    #@+node:ekr.20210611084045.64: *3* sendAsyncOutput
-    def sendAsyncOutput(self, p_package):
-        if "async" not in p_package:
-            print('[sendAsyncOutput] Error async member missing in package parameter')
-            print(json.dumps(p_package, separators=(',', ':')), flush=True)
-            return
-        if self.loop:
-            self.loop.create_task(self.asyncOutput(
-                json.dumps(p_package, separators=(',', ':'))))
-        else:
-            print('[sendAsyncOutput] Error loop not ready' +
-                  json.dumps(p_package, separators=(',', ':')))
-
-    #@+node:ekr.20210611084045.65: *3* set_ask_result
-    def set_ask_result(self, p_result):
-        '''Got the result to an asked question/warning from client'''
-        self.g.app.externalFilesController.integResult(p_result)
-        return self.sendLeoBridgePackage()  # Just send empty as 'ok'
-
-    #@+node:ekr.20210611084045.66: *3* set_config
-    def set_config(self, p_config):
-        '''Got leoInteg's config from client'''
-        self.leoIntegConfig = p_config
-        return self.sendLeoBridgePackage()  # Just send empty as 'ok'
-
-    #@+node:ekr.20210611084045.67: *3* logSignon
-    def logSignon(self):
-        '''Simulate the Initial Leo Log Entry'''
-        if self.loop:
-            self.g.app.computeSignon()
-            self.g.es(str(self.g.app.signon))
-            self.g.es(str(self.g.app.signon1))
-        else:
-            print('no loop in logSignon', flush=True)
-
-    #@+node:ekr.20210611084045.68: *3* JSON Output Functions
-    #@+node:ekr.20210611084045.69: *4* setActionId
-    def setActionId(self, p_id):
-        self.currentActionId = p_id
-
-    #@+node:ekr.20210611084045.70: *4* asyncOutput
-    async def asyncOutput(self, p_json):
-        '''Output json string to the websocket'''
-        if self.webSocket:
-            await self.webSocket.send(p_json)
-        else:
-            print("websocket not ready yet", flush=True)
-
-    #@+node:ekr.20210611084045.71: *4* sendLeoBridgePackage
-    def sendLeoBridgePackage(self, p_package={}):
-        p_package["id"] = self.currentActionId
-        return(json.dumps(p_package, separators=(',', ':')))  # send as json
-
-    #@+node:ekr.20210611084045.72: *4* _outputError
-    def _outputError(self, p_message="Unknown Error"):
-        # Output to this server's running console
-        print("ERROR: " + p_message, flush=True)
-        w_package = {"id": self.currentActionId}
-        w_package["error"] = p_message
-        return p_message
-
-    #@+node:ekr.20210611084045.73: *4* _outputBodyData
-    def _outputBodyData(self, p_bodyText=""):
-        return self.sendLeoBridgePackage({"body": p_bodyText})
-
-    #@+node:ekr.20210611084045.74: *4* _outputSelectionData
-    def _outputSelectionData(self, p_bodySelection):
-        return self.sendLeoBridgePackage({"bodySelection": p_bodySelection})
-
-    #@+node:ekr.20210611084045.75: *4* _outputPNode
-    def _outputPNode(self, p_node=False):
-        if p_node:
-            # Single node, singular
-            return self.sendLeoBridgePackage({"node": self._p_to_ap(p_node)})
-        else:
-            return self.sendLeoBridgePackage({"node": None})
-
-    #@+node:ekr.20210611084045.76: *4* _outputPNodes
-    def _outputPNodes(self, p_pList):
-        w_apList = []
-        for p in p_pList:
-            w_apList.append(self._p_to_ap(p))
-        # Multiple nodes, plural
-        return self.sendLeoBridgePackage({"children": w_apList})
-
-    #@+node:ekr.20210611084045.77: *3* es
-    def es(self, * args, **keys):
-        '''Output to the Log Pane'''
-        d = {
-            'color': None,
-            'commas': False,
-            'newline': True,
-            'spaces': True,
-            'tabName': 'Log',
-            'nodeLink': None,
-        }
-        d = self.g.doKeywordArgs(keys, d)
-        s = self.g.translateArgs(args, d)
-        w_package = {"async": "log", "log": s}
-        self.sendAsyncOutput(w_package)
-    #@+node:ekr.20210611084045.78: *3* initConnection
-    def initConnection(self, p_webSocket):
-        self.webSocket = p_webSocket
-        self.loop = asyncio.get_event_loop()
-
-    #@+node:ekr.20210611084045.79: *3* _get_commander_method
-    def _get_commander_method(self, p_command):
-        """ Return the given method (p_command) in the Commands class or subcommanders."""
-        # self.g.trace(p_command)
+        self.g = g = self.bridge.globals()
+        self.dummy_c = g.app.newCommander(fileName=None)  # To inspect commands
+        self.bad_commands_list = self._bad_commands(self.dummy_c)
         #
-        # First, try the commands class.
-        w_func = getattr(self.commander, p_command, None)
-        if w_func:
-            return w_func
-        #
-        # Search all subcommanders for the method.
-        table = (  # This table comes from c.initObjectIvars.
-            'abbrevCommands',
-            'bufferCommands',
-            'chapterCommands',
-            'controlCommands',
-            'convertCommands',
-            'debugCommands',
-            'editCommands',
-            'editFileCommands',
-            'evalController',
-            'gotoCommands',
-            'helpCommands',
-            'keyHandler',
-            'keyHandlerCommands',
-            'killBufferCommands',
-            'leoCommands',
-            'leoTestManager',
-            'macroCommands',
-            'miniBufferWidget',
-            'printingController',
-            'queryReplaceCommands',
-            'rectangleCommands',
-            'searchCommands',
-            'spellCommands',
-            'vimCommands',  # Not likely to be useful.
-        )
-        for ivar in table:
-            subcommander = getattr(self.commander, ivar, None)
-            if subcommander:
-                w_func = getattr(subcommander, p_command, None)
-                if w_func:
-                    ### self.g.trace(f"Found c.{ivar}.{p_command}")
-                    return w_func
-            # else:
-                # self.g.trace(f"Not Found: c.{ivar}") # Should never happen.
-        return None
-
-    #@+node:ekr.20210611084045.80: *3* leoCommand
-    def leoCommand(self, p_command, param):
-        '''
-        Generic call to a method in Leo's Commands class or any subcommander class.
-
-        The param["ap"] position is to be selected before having the command run,
-        while the param["keep"] parameter specifies wether the original position
-        should be re-selected afterward.
-
-        The whole of those operations is to be undoable as one undo step.
-
-        command: a method name (a string).
-        param["ap"]: an archived position.
-        param["keep"]: preserve the current selection, if possible.
-        '''
-        w_keepSelection = False  # Set default, optional component of param
-        if "keep" in param:
-            w_keepSelection = param["keep"]
-
-        w_ap = param["ap"]  # At least node parameter is present
-        if not w_ap:
-            return self._outputError(f"Error in {p_command}: no param ap")
-
-        w_p = self._ap_to_p(w_ap)
-        if not w_p:
-            return self._outputError(f"Error in {p_command}: no w_p position found")
-
-        w_func = self._get_commander_method(p_command)
-        if not w_func:
-            return self._outputError(f"Error in {p_command}: no method found")
-
-        if w_p == self.commander.p:
-            w_func(event=None)
-        else:
-            w_oldPosition = self.commander.p
-            self.commander.selectPosition(w_p)
-            w_func(event=None)
-            if w_keepSelection and self.commander.positionExists(w_oldPosition):
-                self.commander.selectPosition(w_oldPosition)
-
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.81: *3* Leo Documents
-    #@+node:ekr.20210611084045.82: *4* get_all_open_commanders
-    def get_all_open_commanders(self, param):
-        '''Return array of opened file path/names to be used as openFile parameters to switch files'''
-        w_files = []
-        for w_commander in self.g.app.commanders():
-            if not w_commander.closed:
-                w_isSelected = False
-                w_isChanged = w_commander.changed
-                if self.commander == w_commander:
-                    w_isSelected = True
-                w_entry = {
-                    "name": w_commander.mFileName,
-                    "changed": w_isChanged,
-                    "selected": w_isSelected
-                }
-                w_files.append(w_entry)
-
-        return self.sendLeoBridgePackage({"files": w_files})
-
-    #@+node:ekr.20210611084045.83: *4* get_ui_states
-    def get_ui_states(self, param):
+        # Complete the initialization, as in LeoApp.initApp.
+        g.app.idleTimeManager = leoApp.IdleTimeManager()
+        g.app.idleTimeManager.start()
+        g.app.externalFilesController = leoExternalFiles.ExternalFilesController(None)
+        t2 = time.process_time()
+        print(f"LeoServer: init leoBridge in {t2-t1:4.2} sec.")
+    #@+node:ekr.20210211084004.1: *3* server:public commands
+    #@+node:ekr.20210202193709.1: *4* server:button commands
+    # These will fail unless the open_file inits c.theScriptingController.
+    #@+node:ekr.20210207051720.1: *5* _check_button_command
+    def _check_button_command(self, tag):  # pragma: no cover (no scripting controller)
         """
-        Gets the currently opened file's general states for UI enabled/disabled states
-        such as undo available, file changed/unchanged
+        Check that a button command is possible.
+        Raise ServerError if not. Otherwise, return sc.buttonsDict.
         """
-        w_states = {}
-        if self.commander:
-            try:
-                # 'dirty/changed' member
-                w_states["changed"] = self.commander.changed
-                w_states["canUndo"] = self.commander.canUndo()
-                w_states["canRedo"] = self.commander.canRedo()
-                w_states["canDemote"] = self.commander.canDemote()
-                w_states["canPromote"] = self.commander.canPromote()
-                w_states["canDehoist"] = self.commander.canDehoist()
-
-            except Exception as e:
-                self.g.trace('Error while getting states')
-                print("Error while getting states", flush=True)
-                print(str(e), flush=True)
-        else:
-            w_states["changed"] = False
-            w_states["canUndo"] = False
-            w_states["canRedo"] = False
-            w_states["canDemote"] = False
-            w_states["canPromote"] = False
-            w_states["canDehoist"] = False
-
-        return self.sendLeoBridgePackage({"states": w_states})
-
-    #@+node:ekr.20210611084045.84: *4* set_opened_file
-    def set_opened_file(self, param):
-        '''Choose the new active commander from array of opened file path/names by numeric index'''
-        w_openedCommanders = []
-
-        for w_commander in self.g.app.commanders():
-            if not w_commander.closed:
-                w_openedCommanders.append(w_commander)
-
-        w_index = param['index']  # index in param
-
-        if w_openedCommanders[w_index]:
-            self.commander = w_openedCommanders[w_index]
-
-        if self.commander:
-            self.commander.closed = False
-            self._create_gnx_to_vnode()
-            w_result = {"total": self._getTotalOpened(), "filename": self.commander.fileName(),
-                        "node": self._p_to_ap(self.commander.p)}
-            # maybe needed for frame wrapper
-            self.commander.selectPosition(self.commander.p)
-            return self.sendLeoBridgePackage(w_result)
-        else:
-            return self._outputError('Error in setOpenedFile')
-
-    #@+node:ekr.20210611084045.85: *4* open_file
-    def open_file(self, param):
+        c = self._check_c()
+        sc = getattr(c, "theScriptingController", None)
+        if not sc:
+            # This will happen unless mod_scripting is loaded!
+            raise ServerError(f"{tag}: no scripting controller")
+        return sc.buttonsDict
+    #@+node:ekr.20210202183724.4: *5* server.click_button
+    def click_button(self, package):  # pragma: no cover (no scripting controller)
+        """Handles buttons clicked in client from the '@button' panel"""
+        tag = 'click_button'
+        name = package.get("name")
+        if not name:
+            raise ServerError(f"{tag}: no button name given")
+        d = self._check_button_command(tag)
+        button = d.get(name)
+        if not button:
+            raise ServerError(f"{tag}: button {name!r} does not exist")
+        try:
+            button.command()
+        except Exception as e:
+            raise ServerError(f"{tag}: exception clicking button {name!r}: {e}")
+        return self._make_response()
+    #@+node:ekr.20210202183724.2: *5* server.get_buttons
+    def get_buttons(self, package):  # pragma: no cover (no scripting controller)
+        """Gets the currently opened file's @buttons list"""
+        d = self._check_button_command('get_buttons')
+        return self._make_response({
+            "buttons": sorted(list(d.get.keys()))
+        })
+    #@+node:ekr.20210202183724.3: *5* server.remove_button
+    def remove_button(self, package):  # pragma: no cover (no scripting controller)
+        """Remove button by name."""
+        tag = 'remove_button'
+        name = package.get("name")
+        if not name:
+            raise ServerError(f"{tag}: no button name given")
+        d = self._check_button_command(tag)
+        if name not in d:
+            raise ServerError(f"{tag}: button {name!r} does not exist")
+        try:
+            del d [name]
+        except Exception as e:
+            raise ServerError(f"{tag}: exception removing button {name!r}: {e}")
+        return self._make_response({
+            "buttons": sorted(list(d.get.keys()))
+        })
+    #@+node:ekr.20210202193642.1: *4* server:file commands
+    #@+node:ekr.20210202110128.57: *5* server.open_file
+    def open_file(self, package):
         """
-        Open a leo file via leoBridge controller, or create a new document if empty string.
-        Returns an object that contains a 'opened' member.
+        Open a leo file with the given filename.
+        Create a new document if no name.
         """
-        w_found = False
-        w_filename = param.get('filename')  # Optional.
-        # If not empty string (asking for New file) then check if already opened
-        if w_filename:
-            for w_commander in self.g.app.commanders():
-                if w_commander.fileName() == w_filename:
-                    w_found = True
-                    self.commander = w_commander
-
-        if not w_found:
-            self.commander = self.bridge.openLeoFile(
-                w_filename)  # create self.commander
-            self.commander.findCommands.ftm = StringFindTabManager(
-                self.commander)
-
-        # Leo at this point has done this too: g.app.windowList.append(c.frame)
-        # and so, now, app.commanders() yields this: return [f.c for f in g.app.windowList]
-
-        if self.commander:
-            self.commander.closed = False
-            if not w_found:
-                # is new so also replace wrapper
-                self.commander.frame.body.wrapper = IntegTextWrapper(
-                    self.commander, "integBody", self.g)
-                self.commander.selectPosition(self.commander.p)
-
-            self._create_gnx_to_vnode()
-            w_result = {"total": self._getTotalOpened(), "filename": self.commander.fileName(),
-                        "node": self._p_to_ap(self.commander.p)}
-            return self.sendLeoBridgePackage(w_result)
-        else:
-            return self._outputError('Error in openFile')
-
-    #@+node:ekr.20210611084045.86: *4* open_files
-    def open_files(self, param):
-        """
-        Opens an array of leo files
-        Returns an object that contains the last 'opened' member.
-        """
-        w_files = []
-        if "files" in param:
-            w_files = param["files"]
-
-        for i_file in w_files:
-            self.commander = None
-            w_found = False
-            # If not empty string (asking for New file) then check if already opened
-            if i_file:
-                for w_commander in self.g.app.commanders():
-                    if w_commander.fileName() == i_file:
-                        w_found = True
-                        self.commander = w_commander
-
-            if not w_found:
-                if os.path.isfile(i_file):
-                    self.commander = self.bridge.openLeoFile(
-                        i_file)  # create self.commander
-                    self.commander.findCommands.ftm = StringFindTabManager(
-                        self.commander)
-            if self.commander:
-                self.commander.closed = False
-                self.commander.frame.body.wrapper = IntegTextWrapper(
-                    self.commander, "integBody", self.g)
-                self.commander.selectPosition(self.commander.p)
-
-        # Done with the last one, it's now the selected commander. Check again just in case.
-        if self.commander:
-            self._create_gnx_to_vnode()
-            w_result = {"total": self._getTotalOpened(), "filename": self.commander.fileName(),
-                        "node": self._p_to_ap(self.commander.p)}
-            return self.sendLeoBridgePackage(w_result)
-        else:
-            return self._outputError('Error in openFiles')
-
-    #@+node:ekr.20210611084045.87: *4* close_file
-    def close_file(self, param):
-        """
-        Closes a leo file. A file can then be opened with "openFile".
-        Returns a 'total' member in the package if close is successful
-
-        """
-        if self.commander:
-            # First, revert to prevent asking user.
-            if param["forced"] and self.commander.changed:
-                self.commander.revert()
-            # Then, if still possible, close it.
-            if param["forced"] or not self.commander.changed:
-                self.commander.closed = True
-                self.commander.close()
-            else:
-                # Cannot close, ask to save, ignore or cancel
-                return self.sendLeoBridgePackage()
-
-        # Switch commanders to first available
-        w_total = self._getTotalOpened()
-        if w_total:
-            self.commander = self._getFirstOpenedCommander()
-        else:
-            self.commander = None
-
-        if self.commander:
-            self._create_gnx_to_vnode()
-            w_result = {"total": self._getTotalOpened(), "filename": self.commander.fileName(),
-                        "node": self._p_to_ap(self.commander.p)}
-        else:
-            w_result = {"total": 0}
-
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.88: *4* save_file
-    def save_file(self, param):
-        '''Saves the leo file. New or dirty derived files are rewritten'''
-        if self.commander:
-            try:
-                if "name" in param:
-                    self.commander.save(fileName=param['name'])
-                else:
-                    self.commander.save()
-            except Exception as e:
-                self.g.trace('Error while saving')
-                print("Error while saving", param['name'], flush=True)
-                print(str(e),  param['name'],  flush=True)
-
-        return self.sendLeoBridgePackage()  # Just send empty as 'ok'
-
-    #@+node:ekr.20210611084045.89: *4* import_any_file
-    def import_any_file(self, param):
-        """
-        Import file(s) from array of file names
-        """
-        c = self.commander
-        g = self.g
-        ic = c.importCommands
-        names = param.get('filenames')
-        if names:
-            g.chdir(names[0])
-        if not names:
-            return self._outputError("Error in import_any_file no filenames found")
-
-        # New in Leo 4.9: choose the type of import based on the extension.
-        derived = [z for z in names if c.looksLikeDerivedFile(z)]
-        others = [z for z in names if z not in derived]
-        if derived:
-            ic.importDerivedFiles(parent=c.p, paths=derived)
-        for fn in others:
-            junk, ext = g.os_path_splitext(fn)
-            ext = ext.lower()  # #1522
-            if ext.startswith('.'):
-                ext = ext[1:]
-            if ext == 'csv':
-                ic.importMindMap([fn])
-            elif ext in ('cw', 'cweb'):
-                ic.importWebCommand([fn], "cweb")
-            # Not useful. Use @auto x.json instead.
-            # elif ext == 'json':
-                # ic.importJSON([fn])
-            elif fn.endswith('mm.html'):
-                ic.importFreeMind([fn])
-            elif ext in ('nw', 'noweb'):
-                ic.importWebCommand([fn], "noweb")
-            elif ext == 'more':
-                # (Félix) leoImport Should be on c?
-                c.leoImport.MORE_Importer(c).import_file(fn)  # #1522.
-            elif ext == 'txt':
-                # (Félix) import_txt_file Should be on c?
-                # #1522: Create an @edit node.
-                c.import_txt_file(c, fn)
-            else:
-                # Make *sure* that parent.b is empty.
-                last = c.lastTopLevel()
-                parent = last.insertAfter()
-                parent.v.h = 'Imported Files'
-                ic.importFilesCommand(
-                    files=[fn],
-                    parent=parent,
-                    treeType='@auto',  # was '@clean'
-                    # Experimental: attempt to use permissive section ref logic.
-                )
-        return self.sendLeoBridgePackage()  # Just send empty as 'ok'
-
-    #@+node:ekr.20210611084045.90: *3* Search
-    #@+node:ekr.20210611084045.91: *4* get_search_settings
-    def get_search_settings(self, param):
-        """
-        Gets search options
-        """
-        w_result = self.commander.findCommands.ftm.get_settings()
-        return self.sendLeoBridgePackage({"searchSettings": w_result.__dict__})
-
-    #@+node:ekr.20210611084045.92: *4* set_search_settings
-    def set_search_settings(self, param):
-        """
-        Sets search options. Init widgets and ivars from param.searchSettings
-        """
-        c = self.commander
-        find = c.findCommands
-        ftm = c.findCommands.ftm
-        searchSettings = param['searchSettings']
-        # Find/change text boxes.
-        table = (
-            ('find_findbox', 'find_text', ''),
-            ('find_replacebox', 'change_text', ''),
-        )
-        for widget_ivar, setting_name, default in table:
-            w = getattr(ftm, widget_ivar)
-            s = searchSettings.get(setting_name) or default
-            w.clear()
-            w.insert(s)
-        # Check boxes.
-        table = (
-            ('ignore_case', 'check_box_ignore_case'),
-            ('mark_changes', 'check_box_mark_changes'),
-            ('mark_finds', 'check_box_mark_finds'),
-            ('pattern_match', 'check_box_regexp'),
-            ('search_body', 'check_box_search_body'),
-            ('search_headline', 'check_box_search_headline'),
-            ('whole_word', 'check_box_whole_word'),
-        )
-        for setting_name, widget_ivar in table:
-            w = getattr(ftm, widget_ivar)
-            val = searchSettings.get(setting_name)
-            setattr(find, setting_name, val)
-            if val != w.isChecked():
-                w.toggle()
-        # Radio buttons
-        table = (
-            ('node_only', 'node_only', 'radio_button_node_only'),
-            ('entire_outline', None, 'radio_button_entire_outline'),
-            ('suboutline_only', 'suboutline_only', 'radio_button_suboutline_only'),
-        )
-        for setting_name, ivar, widget_ivar in table:
-            w = getattr(ftm, widget_ivar)
-            val = searchSettings.get(setting_name, False)
-            if ivar is not None:
-                assert hasattr(find, setting_name), setting_name
-                setattr(find, setting_name, val)
-                if val != w.isChecked():
-                    w.toggle()
-        # Ensure one radio button is set.
-        w = ftm.radio_button_entire_outline
-        if not searchSettings.get('node_only', False) and not searchSettings.get('suboutline_only', False):
-            setattr(find, 'entire_outline', True)
-            if not w.isChecked():
-                w.toggle()
-        else:
-            setattr(find, 'entire_outline', False)
-            if w.isChecked():
-                w.toggle()
-
-        # Confirm by sending back the settings to leointeg
-        w_result = ftm.get_settings()
-        return self.sendLeoBridgePackage({"searchSettings": w_result.__dict__})
-
-    #@+node:ekr.20210611084045.93: *4* find_all
-    def find_all(self, param):
-        """Run Leo's find all command and return results."""
-        c = self.commander
+        found, tag = False, 'open_file'
+        filename = package.get('filename')  # Optional.
+        if filename:
+            for c in g.app.commanders():
+                if c.fileName() == filename:
+                    found = True
+        if not found:
+            c = self.bridge.openLeoFile(filename)
+            c.findCommands.ftm = StringFindTabManager(c)
+        if not c:  # pragma: no cover
+            raise ServerError(f"{tag}: bridge did not open {filename!r}")
+        if not c.frame.body.wrapper:  # pragma: no cover
+            raise ServerError(f"{tag}: no wrapper")
+        # Assign self.c
+        self.c = c
+        c.selectPosition(c.rootPosition())  # Required.
+        # Check the outline!
+        self._check_outline(c)
+        if self.log_flag:  # pragma: no cover
+            self._dump_outline(c)
+        return self._make_response()
+    #@+node:ekr.20210202110128.58: *5* server.close_file
+    def close_file(self, package):
+        """Closes an outline opened with open_file."""
+        c = self._check_c()
+        # Close the outline, even if it is dirty!
+        c.clearChanged()
+        c.close()
+        # Select the first open outline, if any.
+        commanders = g.app.commanders()
+        self.c = commanders and commanders[0] or None
+        # Return a response describing self.c, not the closed outline.
+        return self._make_response()
+    #@+node:ekr.20210202183724.1: *5* server.save_file
+    def save_file(self, package):  # pragma: no cover (too dangerous).
+        """Save the leo outline."""
+        c = self._check_c()
+        c.save()
+        return self._make_response()
+    #@+node:ekr.20210212092848.1: *4* server:find commands
+    #@+node:ekr.20210212094817.1: *5* server._get_find_settings
+    def _get_find_settings(self, c):
+        """Return a g.Bunch containing the present find settings settings."""
+        return c.findCommands.ftm.get_settings()
+        ###
+            # # For now, return EKR defaults.
+            # return g.Bunch(
+                # find_text=None, change_text=None,
+                # search_body=True, search_headline=True,
+                # ignore_case=True, pattern_match=False, whole_word=True,
+                # mark_changes=False, mark_finds=False,
+                # node_only=False, suboutline_only=False,
+                # entry_focus=None,
+            # )  
+    #@+node:ekr.20210212092854.1: *5* server.find_all
+    def find_all(self, package):
+        """Run Leo's find-all command and return results."""
+        tag = 'find_all'
+        c = self._check_c()
         fc = c.findCommands
-        settings = fc.ftm.get_settings()
-        result = fc.do_find_all(settings)
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        w_result = {"found": result,
-                    "focus": focus, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.94: *4* find_next
-    def find_next(self, param):
-        """Run Leo's find-next command and return results."""
-        c = self.commander
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        answer = fc.do_find_all(settings)
+        if self.log_flag:  # pragma: no cover
+            g.printObj(answer, tag=f"{tag}: answer")
+        return self._make_response({"answer": answer})
+    #@+node:ekr.20210221042145.1: *5* server.change_all
+    def change_all(self, package):
+        """Run Leo's change-all command and return results."""
+        tag = 'change_all'
+        c = self._check_c()
         fc = c.findCommands
-        settings = fc.ftm.get_settings()
-        p, pos, newpos = fc.do_find_next(settings)
-        found = True
-        if not p:
-            found = False
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        w_result = {"found": found, "pos": pos, "newpos": newpos,
-                    "focus": focus, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.95: *4* find_previous
-    def find_previous(self, param):
-        """Run Leo's find-previous command and return results."""
-        c = self.commander
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        change_text = package.get("change_text")
+        if change_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no change text")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        settings.change_text = change_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        answer = fc.do_change_all(settings)
+        if self.log_flag:  # pragma: no cover
+            g.printObj(answer, tag=f"{tag}: answer")
+        return self._make_response({"answer": answer})
+    #@+node:ekr.20210221042406.1: *5* server.change_then_find
+    def change_then_find(self, package):
+        """Run Leo's change-then-find command and return results."""
+        tag = 'change_then_find'
+        c = self._check_c()
         fc = c.findCommands
-        settings = fc.ftm.get_settings()
-        p, pos, newpos = fc.do_find_prev(settings)
-        found = True
-        if not p:
-            found = False
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        w_result = {"found": found, "pos": pos, "newpos": newpos,
-                    "focus": focus, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-
-    #@+node:ekr.20210611084045.96: *4* replace
-    def replace(self, param):
-        """Run Leo's replace command and return results."""
-        c = self.commander
-        fc = c.findCommands
-        settings = fc.ftm.get_settings()
-        fc.change(settings)
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        w_result = {"found": True,
-                    "focus": focus, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.97: *4* replace_then_find
-    def replace_then_find(self, param):
-        """Run Leo's replace then find next command and return results."""
-        c = self.commander
-        fc = c.findCommands
-        settings = fc.ftm.get_settings()
-        result = fc.do_change_then_find(settings)
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        w_result = {"found": result,
-                    "focus": focus, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.98: *4* replace_all
-    def replace_all(self, param):
-        """Run Leo's replace all command and return results."""
-        c = self.commander
-        fc = c.findCommands
-        settings = fc.ftm.get_settings()
-        result = fc.do_change_all(settings)
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        w_result = {"found": result,
-                    "focus": focus, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.99: *4* clone_find_all
-    def clone_find_all(self, param):
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        answer = fc.do_change_then_find(settings)
+        if self.log_flag:  # pragma: no cover
+            g.printObj(answer, tag=f"{tag}: answer")
+        return self._make_response({"answer": answer})
+    #@+node:ekr.20210221042541.1: *5* server.clone_find_all
+    def clone_find_all(self, package):
         """Run Leo's clone-find-all command and return results."""
-        c = self.commander
+        tag = 'clone_find_all'
+        c = self._check_c()
         fc = c.findCommands
-        settings = fc.ftm.get_settings()
-        result = fc.do_clone_find_all(settings)
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        w_result = {"found": result,
-                    "focus": focus, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.100: *4* clone_find_all_flattened
-    def clone_find_all_flattened(self, param):
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        answer = fc.do_clone_find_all(settings)
+        if self.log_flag:  # pragma: no cover
+            g.printObj(answer, tag=f"{tag}: answer")
+        return self._make_response({"answer": answer})
+    #@+node:ekr.20210221042633.1: *5* server.clone_find_all_flattened
+    def clone_find_all_flattened(self, package):
         """Run Leo's clone-find-all-flattened command and return results."""
-        c = self.commander
+        tag = 'clone_find_all_flattened'
+        c = self._check_c()
         fc = c.findCommands
-        settings = fc.ftm.get_settings()
-        result = fc.do_clone_find_all_flattened(settings)
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        w_result = {"found": result,
-                    "focus": focus, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-    #@+node:ekr.20210611084045.101: *4* find_var
-
-    def find_var(self, param):
-        """Run Leo's find-var command and return results."""
-        c = self.commander
-        ### fc = c.findCommands
-        ### settings = fc.ftm.get_settings()
-        # todo : find var implementation
-        print("todo : find var implementation")
-        # result = fc.do_clone_find_all_flattened(settings)
-        ### w = self.g.app.gui.get_focus()
-        ### focus = self.g.app.gui.widget_name(w)
-        w_result = {"node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.102: *4* find_def
-    def find_def(self, param):
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        answer = fc.do_clone_find_all_flattened(settings)
+        if self.log_flag:  # pragma: no cover
+            g.printObj(answer, tag=f"{tag}: answer")
+        return self._make_response({"answer": answer})
+    #@+node:ekr.20210221042719.1: *5* server.clone_find_tag
+    def clone_find_tag(self, package):
+        """Run Leo's clone-find-tag command and return results."""
+        tag = 'clone_find_tag'
+        c = self._check_c()
+        fc = c.findCommands
+        the_tag = package.get("tag")
+        if not the_tag:  # pragma: no cover
+            raise ServerError(f"{tag}: no tag")
+        settings = self._get_find_settings(c)
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        n, p = fc.do_clone_find_tag(settings)
+        if self.log_flag:  # pragma: no cover
+            g.trace("tag: {the_tag} n: {n} p: {p and p.h!r}")
+        return self._make_response({"n": n, "p": p})
+    #@+node:ekr.20210221043043.1: *5* server.find_def
+    def find_def(self, package):
         """Run Leo's find-def command and return results."""
-        c = self.commander
-        ### fc = c.findCommands
-        ### settings = fc.ftm.get_settings()
-        # todo : find def implementation
-        print("todo : find def implementation")
-        # result = fc.do_clone_find_all_flattened(settings)
-        ### w = self.g.app.gui.get_focus()
-        ### focus = self.g.app.gui.widget_name(w)
-        w_result = {"node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
+        tag = 'find_def'
+        c = self._check_c()
+        fc = c.findCommands
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        p, pos, newpos = fc.do_find_def(settings, word=find_text, strict=False)
+        if self.log_flag:  # pragma: no cover
+            g.trace(f"p: {p and p.h!r} pos: {pos!r} newpos {newpos!r}")
+        return self._make_response({"p": p, "pos": pos, "newpos": newpos})
+    #@+node:ekr.20210221042808.1: *5* server.find_next
+    def find_next(self, package):
+        """Run Leo's find-next command and return results."""
+        tag = 'find_next'
+        c = self._check_c()
+        fc = c.findCommands
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        p, pos, newpos = fc.do_find_next(settings)
+        if self.log_flag:  # pragma: no cover
+            g.trace(f"p: {p and p.h!r} pos: {pos!r} newpos {newpos!r}")
+        return self._make_response({"p": p, "pos": pos, "newpos": newpos})
+    #@+node:ekr.20210221042851.1: *5* server.find_previous
+    def find_previous(self, package):
+        """Run Leo's find-previous command and return results."""
+        tag = 'find_previous'
+        c = self._check_c()
+        fc = c.findCommands
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        p, pos, newpos = fc.do_find_prev(settings)
+        if self.log_flag:  # pragma: no cover
+            g.trace(f"p: {p and p.h!r} pos: {pos!r} newpos {newpos!r}")
+        return self._make_response({"p": p, "pos": pos, "newpos": newpos})
+    #@+node:ekr.20210221043134.1: *5* server.find_var
+    def find_var(self, package):
+        """Run Leo's find-var command and return results."""
+        tag = 'find_var'
+        c = self._check_c()
+        fc = c.findCommands
+        find_text = package.get("find_text")
+        if find_text is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no find pattern")
+        settings = self._get_find_settings(c)
+        settings.find_text = find_text
+        if self.log_flag:  # pragma: no cover
+            g.printObj(settings, tag=f"{tag}: settings for {c.shortFileName()}")
+        p, pos, newpos = fc.do_find_var(settings, word=find_text)
+        if self.log_flag:  # pragma: no cover
+            g.trace(f"p: {p and p.h!r} pos: {pos!r} newpos {newpos!r}")
+        return self._make_response({"p": p, "pos": pos, "newpos": newpos})
+    #@+node:ekr.20210221043224.1: *5* server.tag_children
+    def tag_children(self, package):
+        """Run Leo's tag-children command and return results."""
+        # This is not a find command!
+        tag = 'tag_children'
+        c = self._check_c()
+        fc = c.findCommands
+        the_tag = package.get("tag")
+        if the_tag is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no tag")
+        # Unlike find commands, do_tag_children does not use a settings dict.
+        fc.do_tag_children(c.p, the_tag)
+        return self._make_response({})
+    #@+node:ekr.20210202193505.1: *4* server:getter commands
+    #@+node:ekr.20210202110128.55: *5* server.get_all_open_commanders
+    def get_all_open_commanders(self, package):
+        """Return array describing each commander in g.app.commanders()."""
+        files = [
+            {
+                "changed": c.isChanged(),
+                "name": c.fileName(),
+                "selected": c == self.c,
+            } for c in g.app.commanders()
+        ]
+        return self._make_response({"open-commanders": files})
+    #@+node:ekr.20210202110128.71: *5* server.get_all_positions
+    def get_all_positions(self, package):
+        """
+        Return a list of position data for all positions.
+        
+        Useful as a sanity check for debugging.
+        """
+        c = self._check_c()
+        result = [
+            self._get_position_d(p) for p in c.all_positions(copy=False)
+        ]
+        return self._make_response({"position-data-list": result})
+    #@+node:ekr.20210202110128.72: *5* server.get_body & get_body_length
+    def get_body(self, package):
+        """
+        Return p.b, where p is c.p if package["ap"] is missing.
+        
+        Note: There is no need for a separate get_body_length command,
+              because _make_response always adds "body-length": len(p.b)
+        """
+        self._check_c()
+        p = self._get_p(package)
+        # _make_response adds all the cheap redraw data, including "body-length"
+        return self._make_response({"body": p.b})
+        
+    #@+node:ekr.20210202110128.66: *5* server.get_body_states
+    def get_body_states(self, package):
+        """
+        Return body data for p, where p is c.p if package["ap"] is missing.
+        """
+        c = self._check_c()
+        p = self._get_p(package)
+        wrapper = c.frame.body.wrapper
+        
+        def row_col_dict(i):
+            junk, line, col = wrapper.toPythonIndexRowCol(i)
+            return {"line": line, "col": col}
+            
+        # Get the language.
+        aList = g.get_directives_dict_list(p)
+        d = g.scanAtCommentAndAtLanguageDirectives(aList)
+        language = (
+            d and d.get('language')
+            or g.getLanguageFromAncestorAtFileNode(p)
+            or c.config.getLanguage('target-language')
+            or 'plain'
+        )
+        # get values from wrapper if it's the selected node.
+        if c.p.v.gnx == p.v.gnx:
+            active = wrapper.getInsertPoint()
+            start, end = wrapper.getSelectionRange(True)
+            scroll = wrapper.getYScrollPosition()
+        else:  # pragma: no cover
+            active = p.v.insertSpot
+            start = p.v.selectionStart
+            end = p.v.selectionStart + p.v.selectionLength
+            scroll = p.v.scrollBarSpot
+        states = {
+            'language': language.lower(),
+            'selection': {
+                # "gnx": p.v.gnx,  # EKR: Not needed. The reponse will have p.v.gnx.
+                "scroll": scroll,
+                "active": row_col_dict(active),
+                "start": row_col_dict(start),
+                "end": row_col_dict(end),
+            }
+        }
+        return self._make_response({"body-states": states})
+    #@+node:ekr.20210202110128.68: *5* server.get_children
+    def get_children(self, package):
+        """
+        Return the node data for children of p, where p is c.p if package["ap"] is missing."""
+        self._check_c()
+        p = self._get_p(package)
+        return self._make_response({
+            # "children": [self._p_to_ap(child) for child in p.children()]
+            "children": [self._get_position_d(child) for child in p.children()]
+        })
+    #@+node:ekr.20210214154702.1: *5* server.get_focus
+    def get_focus(self, packages):
+        """
+        Return a representation of the focs widget,
+        one of ("body", "tree", "headline", repr(the_widget)).
+        """
+        w = g.app.gui.get_focus()
+        focus = g.app.gui.widget_name(w)
+        return self._make_response({"focus": focus})
+    #@+node:ekr.20210202110128.69: *5* server.get_parent
+    def get_parent(self, package):
+        """Return the node data for the parent of position p, where p is c.p if package["ap"] is missing."""
+        self._check_c()
+        p = self._get_p(package)
+        parent = p.parent()
+        data = self._get_position_d(parent) if parent else None
+        return self._make_response({"parent": data})
+    #@+node:ekr.20210211053955.1: *5* server.get_position_dict
+    def get_position_data_dict(self, package):
+        """
+        Return a dict of postition data for all positions.
+        
+        Useful as a sanity check for debugging.
+        """
+        c = self._check_c()
+        result = {
+            p.v.gnx: self._get_position_d(p)
+                for p in c.all_unique_positions(copy=False)
+        }
+        return self._make_response({"position-data-dict": result})
+    #@+node:ekr.20210211233814.1: *5* server.get_ua
+    def get_ua(self, package):
+        """Return p.v.u, making sure it can be serialized."""
+        self._check_c()
+        p = self._get_p(package)
+        try:
+            ua = {"ua": p.v.u}
+            json.dumps(ua, separators=(',', ':'))
+            response = {"p": p, "ua": p.v.u} 
+        except Exception:  # pragma: no cover
+            response = {"p": p, "bad-ua": repr(p.v.u)} 
+        # _make_response adds all the cheap redraw data.
+        return self._make_response(response)
+    #@+node:ekr.20210206062654.1: *5* server.get_sign_on
+    def get_sign_on(self, package):
+        """Synchronous version of _sign_on"""
+        g.app.computeSignon()
+        signon = []
+        for z in (g.app.signon, g.app.signon1):
+            for z2 in z.split('\n'):
+                signon.append(z2.strip())
+        return self._make_response({"sign-on": "\n".join(signon)})
+    #@+node:ekr.20210202110128.61: *5* server.get_ui_states
+    def get_ui_states(self, package):
+        """
+        Return the enabled/disabled UI states for the open commander, or defaults if None.
+        """
+        c = self._check_c()
+        tag = 'get_ui_states'
+        try:
+            states = {
+                "changed": c and c.changed,
+                "canUndo": c and c.canUndo(),
+                "canRedo": c and c.canRedo(),
+                "canDemote": c and c.canDemote(),
+                "canPromote": c and c.canPromote(),
+                "canDehoist": c and c.canDehoist(),
+            }
+        except Exception as e:  # pragma: no cover
+            raise ServerError(f"{tag}: Exception setting state: {e}")
+        return self._make_response({"states": states})
+    #@+node:ekr.20210202193540.1: *4* server:node commands
+    #@+node:ekr.20210202183724.11: *5* server.clone_node
+    def clone_node(self, package):
+        """
+        Clone the node at position p, where p is c.p if package["ap"] is missing.
+        
+        To clone c.p, use this request:
+        {
+            "action": "execute-leo-command",
+            "leo-command-name": "clone",
+        }
+        """
+        c = self._check_c()
+        p = self._get_p(package)
+        c.selectPosition(p)
+        c.clone()
+        return self._make_response()
+    #@+node:ekr.20210202110128.79: *5* server.contract_node
+    def contract_node(self, package):
+        """
+        Contract the node at position p, where p is c.p if package["ap"] is missing.
+        
+        To contract c.p, use this request:
+        {
+            "action": "execute-leo-command",
+            "leo-command-name": "contract-node",
+        }
+        """
+        self._check_c()
+        p = self._get_p(package)
+        p.contract()
+        return self._make_response()
+    #@+node:ekr.20210202183724.12: *5* server.cut_node
+    def cut_node(self, package):  # pragma: no cover (too dangerous, for now)
+        """
+        Cut the node (and its descendants) at position p, where p is c.p if package["ap"] is missing.
+        
+        To cut c.p, use this request:
+        {
+            "action": "execute-leo-command",
+            "leo-command-name": "cut-node",
+        }
+        """
+        c = self._check_c()
+        p = self._get_p(package)
+        c.selectPosition(p)
+        c.cutOutline()
+        return self._make_response()
+    #@+node:ekr.20210202183724.13: *5* server.delete_node
+    def delete_node(self, package):  # pragma: no cover (too dangerous, for now)
+        """
+        Delete the node (and its descendants) at position p, where p is c.p if package["ap"] is missing.
+        
+        To delete c.p, use this request:
+        {
+            "action": "execute-leo-command",
+            "leo-command-name": "delete-node",
+        }
+        """
+        c = self._check_c()
+        p = self._get_p(package)
+        c.selectPosition(p)
+        c.deleteOutline()  # Handles undo.
+        return self._make_response()
+    #@+node:ekr.20210202110128.78: *5* server.expand_node
+    def expand_node(self, package):
+        """
+        Expand the node at position p, where p is c.p if package["ap"] is missing.
+        
+        To expand c.p, use this request:
+        {
+            "action": "execute-leo-command",
+            "leo-command-name": "expand-node",
+        }
+        """
+        self._check_c()
+        p = self._get_p(package)
+        p.expand()
+        return self._make_response()
+    #@+node:ekr.20210202183724.15: *5* server.insert_node
+    def insert_node(self, package):
+        """
+        Insert a new node at position p, where p is c.p if package["ap"] is missing.
 
-    #@+node:ekr.20210611084045.103: *4* goto_global_line
-    def goto_global_line(self, param):
-        """Run Leo's goto-global-line command and return results."""
-        c = self.commander
-        junk_p, junk_offset, found = c.gotoCommands.find_file_line(
-            n=int(param['line']))
-        w_result = {"found": found, "node": self._p_to_ap(c.p)}
-        return self.sendLeoBridgePackage(w_result)
-
-    #@+node:ekr.20210611084045.104: *3* At Buttons
-    #@+node:ekr.20210611084045.105: *4* get_buttons
-    def get_buttons(self, param):
-        '''Gets the currently opened file's @buttons list'''
-        w_buttons = []
-        if self.commander and self.commander.theScriptingController and self.commander.theScriptingController.buttonsDict:
-            w_dict = self.commander.theScriptingController.buttonsDict
-            for w_key in w_dict:
-                w_entry = {"name": w_dict[w_key], "index": str(w_key)}
-                w_buttons.append(w_entry)
-        return self.sendLeoBridgePackage({"buttons": w_buttons})
-
-    #@+node:ekr.20210611084045.106: *4* remove_button
-    def remove_button(self, param):
-        '''Removes an entry from the buttonsDict by index string'''
-        w_index = param['index']
-        w_dict = self.commander.theScriptingController.buttonsDict
-        w_key = None
-        for i_key in w_dict:
-            if(str(i_key) == w_index):
-                w_key = i_key
-        if w_key:
-            del(w_dict[w_key])  # delete object member
-        # return selected node when done
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.107: *4* click_button
-    def click_button(self, param):
-        '''Handles buttons clicked in client from the '@button' panel'''
-        w_index = param['index']
-        w_dict = self.commander.theScriptingController.buttonsDict
-        w_button = None
-        for i_key in w_dict:
-            if(str(i_key) == w_index):
-                w_button = i_key
-        if w_button:
-            w_button.command()  # run clicked button command
-        # return selected node when done
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.108: *3* Minibuffer
-    #@+node:ekr.20210611084045.109: *4* get_all_leo_commands
-    def get_all_leo_commands(self, param):
+        This node has 'newHeadline' as its headline.
+        
+        To insert a new node at c.p (with the default headline), use this request:
+        {
+            "action": "execute-leo-command",
+            "leo-command-name": "insert-node",
+        }
+        
+        Use the 'set_headline' method to undoably set any node's headlines.
+        """
+        c = self._check_c()
+        p = self._get_p(package)
+        c.selectPosition(p)
+        c.insertHeadline()  # Handles undo, sets c.p
+        return self._make_response()
+    #@+node:ekr.20210202110128.64: *5* server.page_down
+    def page_down(self, package):
+        """
+        Selects a node "n" steps down in the tree to simulate page down.
+        """
+        c = self._check_c()
+        n = package.get("n", 3)
+        for z in range(n):
+            c.selectVisNext()
+        return self._make_response()
+    #@+node:ekr.20210202110128.63: *5* server.page_up
+    def page_up(self, package):
+        """
+        Selects a node "N" steps up in the tree to simulate page up.
+        """
+        c = self._check_c()
+        n = package.get("n", 3)
+        for z in range(n):
+            c.selectVisBack()
+        return self._make_response()
+    #@+node:ekr.20210202183724.17: *5* server.redo
+    def redo(self, package):
+        """Undo last un-doable operation"""
+        c = self._check_c()
+        u = c.undoer
+        if u.canRedo():
+            u.redo()
+        return self._make_response()
+    #@+node:ekr.20210202110128.74: *5* server.set_body
+    def set_body(self, package):
+        """
+        Undoably set p.b, where p is c.p if package["ap"] is missing.
+        """
+        tag = 'set_body'
+        c = self._check_c()
+        p = self._get_p(package)
+        u, wrapper = c.undoer, c.frame.body.wrapper
+        body = package.get('body')
+        if body is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no body given")
+        bunch = u.beforeChangeNodeContents(p)
+        p.v.setBodyString(body)
+        u.afterChangeNodeContents(p, "Body Text", bunch)
+        if c.p == p:
+            wrapper.setAllText(body)
+        if not self.c.isChanged():  # pragma: no cover
+            c.setChanged()
+        if not p.v.isDirty():  # pragma: no cover
+            p.setDirty()
+        return self._make_response()
+    #@+node:ekr.20210202110128.77: *5* server.set_current_position
+    def set_current_position(self, package):
+        """Select position p, where p is c.p if package["ap"] is missing."""
+        c = self._check_c()
+        p = self._get_p(package)
+        c.selectPosition(p)
+        return self._make_response()
+    #@+node:ekr.20210202110128.76: *5* server.set_headline
+    def set_headline(self, package):
+        """
+        Undoably set p.h, where p is c.p if package["ap"] is missing.
+        """
+        tag = 'set_headline'
+        c = self._check_c()
+        p = self._get_p(package)
+        u = c.undoer
+        h = package.get('headline')
+        if not h:  # pragma: no cover
+            raise ServerError(f"{tag}: no headline")
+        bunch = u.beforeChangeNodeContents(p)
+        p.h = h
+        u.afterChangeNodeContents(p, 'Change Headline', bunch)
+        return self._make_response()
+    #@+node:ekr.20210202110128.75: *5* server.set_selection
+    def set_selection(self, package):
+        """
+        Set the selection range for p.b, where p is c.p if package["ap"] is missing.
+        
+        Set the selection in the wrapper if p == c.p
+        
+        Package has these keys:
+            
+        - "ap":     An archived position for position p.
+        - "start":  The start of the selection.
+        - "end":    The end of the selection.
+        - "insert": The insert point. Must be either start or end.
+        - "scroll": An optional scroll position.
+        """
+        c = self._check_c()
+        p = self._get_p(package)  # Will raise ServerError if p does not exist.
+        v = p.v
+        wrapper = c.frame.body.wrapper
+        start = package.get('start', 0)
+        end = package.get('end', 0)
+        insert = package.get('insert', 0)
+        scroll = package.get('scroll', 0)
+        if p == c.p:
+            wrapper.setSelectionRange(start, end, insert)
+            wrapper.setYScrollPosition(scroll)
+        # Always set vnode attrs.
+        v.scrollBarSpot = scroll
+        v.insertSpot = insert
+        v.selectionStart = start
+        v.selectionLength = abs(start - end)
+        return self._make_response()
+    #@+node:ekr.20210202183724.10: *5* server.toggle_mark
+    def toggle_mark(self, package):
+        """
+        Toggle the mark at position p, where p is c.p if package["ap"] is missing.
+        
+        To *toggle* the mark of c.p, use this request:
+        {
+            "action": "execute-leo-command",
+            "leo-command-name": "toggle-mark",
+        }
+        """
+        self._check_c()
+        p = self._get_p(package)
+        if p.isMarked():
+            p.clearMarked()
+        else:
+            p.setMarked()
+        return self._make_response()
+    #@+node:ekr.20210202183724.16: *5* server.undo
+    def undo(self, package):
+        """Undo last un-doable operation"""
+        c = self._check_c()
+        u = c.undoer
+        if u.canUndo():
+            u.undo()
+        # Félix: Caller can get focus using other calls.
+        return self._make_response()
+    #@+node:ekr.20210205102806.1: *4* server:server commands
+    #@+node:ekr.20210205102818.1: *5* server.error
+    def error(self, package):
+        """For unit testing. Raise ServerError"""
+        raise ServerError("error called")
+    #@+node:ekr.20210202183724.5: *5* server.get_all_leo_commands & helper
+    def get_all_leo_commands(self, package):
         """Return a list of all Leo commands that make sense in leoInteg."""
-        c = self.commander
+        tag = 'get_all_leo_commands'
+        c = self.dummy_c  # Use the dummy commander.
         d = c.commandsDict  # keys are command names, values are functions.
-        bad_names = self._bad_commands()  # #92.
+        bad_names = self._bad_commands(c)  # #92.
         good_names = self._good_commands()
         duplicates = set(bad_names).intersection(set(good_names))
-        if duplicates:
-            print('duplicate command names...', flush=True)
+        if duplicates:  # pragma: no cover
+            print('duplicate command names...')
             for z in sorted(duplicates):
                 print(z)
         result = []
         for command_name in sorted(d):
             func = d.get(command_name)
-            if not func:
-                print('no func:', command_name, flush=True)
+            if not func:  # pragma: no cover
+                print('no func:', command_name)
                 continue
             if command_name in bad_names:  # #92.
                 continue
             # Prefer func.__func_name__ to func.__name__: Leo's decorators change func.__name__!
             func_name = getattr(func, '__func_name__', func.__name__)
-            if not func_name:
-                print('no name', command_name, flush=True)
+            if not func_name:  # pragma: no cover
+                print('no name', command_name)
                 continue
             doc = func.__doc__ or ''
             result.append({
-                "label": command_name,
+                "command-name": command_name,
                 "func":  func_name,
                 "detail": doc,
             })
-            # This shows up in the bridge log.
-            # print(f"__doc__: {len(doc):4} {command_name:40} {func_name} ", flush=True)
-            # print(f"{func_name} ", flush=True)
-
-        return self.sendLeoBridgePackage({"commands": result})
-
-    #@+node:ekr.20210611084045.110: *4* _bad_commands
-    def _bad_commands(self):
+        if self.log_flag:  # pragma: no cover
+            print(f"\n{tag}: {len(result)} leo commands\n")
+            g.printObj([z.get("command-name") for z in result], tag=tag)
+        return self._make_response({"commands": result})
+    #@+node:ekr.20210202183724.6: *6* server._bad_commands
+    def _bad_commands(self, c):
         """Return the list of Leo's command names that leoInteg should ignore."""
-        c = self.commander
-        bad = []
         d = c.commandsDict  # keys are command names, values are functions.
+        bad = []
         #
         # First, remove @button, @command and vim commands.
         for command_name in sorted(d):
@@ -2187,7 +1426,7 @@ class LeoBridgeIntegController:
         result = list(sorted(bad))
         return result
 
-    #@+node:ekr.20210611084045.111: *4* _good_commands
+    #@+node:ekr.20210202183724.7: *6* server._good_commands
     def _good_commands(self):
         """Defined commands that definitely should be included in leoInteg."""
         good_list = [
@@ -2406,8 +1645,8 @@ class LeoBridgeIntegController:
             'write-missing-at-file-nodes',
             'write-outline-only',
 
-            'clone-find-all',  # Should be overridden by leointeg
-            'clone-find-all-flattened',   # Should be overridden by leointeg
+            'clone-find-all',
+            'clone-find-all-flattened',
             'clone-find-all-flattened-marked',
             'clone-find-all-marked',
             'clone-find-parents',
@@ -2436,7 +1675,7 @@ class LeoBridgeIntegController:
 
             'pdb',
 
-            'redo',  # Should be overridden by leointeg
+            'redo',
             'rst3',
             'run-all-unit-tests-externally',
             'run-all-unit-tests-locally',
@@ -2446,7 +1685,7 @@ class LeoBridgeIntegController:
             'run-selected-unit-tests-locally',
             'run-tests',
 
-            'undo',  # Should be overridden by leointeg
+            'undo',
 
             'xdb',
             # Beautify, blacken, fstringify...
@@ -2569,7 +1808,6 @@ class LeoBridgeIntegController:
             'reformat-body',
             'reformat-paragraph',
             'refresh-from-disk',
-            'reload-all-settings',
             'reload-settings',
             'reload-style-sheets',
             'revert',
@@ -2619,679 +1857,373 @@ class LeoBridgeIntegController:
         ]
         return good_list
 
-    #@+node:ekr.20210611084045.112: *4* _getDocstringForCommand
-    def _getDocstringForCommand(self, command_name):
-        """get docstring for the given command."""
-        func = self._get_commander_method(command_name)
-        docstring = func.__doc__ if func else ''
-        return docstring
-
-    #@+node:ekr.20210611084045.113: *3* Overriden Leo Commands
-    #@+node:ekr.20210611084045.114: *4* mark_node
-    def mark_node(self, param):
-        '''Mark a node, don't select it'''
-        w_ap = param["ap"]
-        if w_ap:
-            w_p = self._ap_to_p(w_ap)
-            if w_p:
-                w_p.setMarked()
-                # return selected node when done (not w_p)
-                return self._outputPNode(self.commander.p)
-            else:
-                return self._outputError("Error in markPNode no w_p node found")
-        else:
-            return self._outputError("Error in markPNode no param node")
-
-    #@+node:ekr.20210611084045.115: *4* unmark_node
-    def unmark_node(self, param):
-        '''Unmark a node, don't select it'''
-        w_ap = param["ap"]
-        if w_ap:
-            w_p = self._ap_to_p(w_ap)
-            if w_p:
-                w_p.clearMarked()
-                # return selected node when done (not w_p)
-                return self._outputPNode(self.commander.p)
-            else:
-                return self._outputError("Error in unmarkPNode no w_p node found")
-        else:
-            return self._outputError("Error in unmarkPNode no param node")
-
-    #@+node:ekr.20210611084045.116: *4* clone_node
-    def clone_node(self, param):
-        '''Clone a node, return it, if it was also the current selection, otherwise try not to select it'''
-        w_ap = param["ap"]
-        if not w_ap:
-            return self._outputError("Error in clonePNode function, no param p_ap")
-        w_p = self._ap_to_p(w_ap)
-        if not w_p:
-            # default empty
-            return self._outputError("Error in clonePNode function, no w_p node found")
-        if w_p == self.commander.p:
-            self.commander.clone()
-        else:
-            oldPosition = self.commander.p
-            self.commander.selectPosition(w_p)
-            self.commander.clone()
-            if self.commander.positionExists(oldPosition):
-                self.commander.selectPosition(oldPosition)
-        # return selected node either ways
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.117: *4* cut_node
-    def cut_node(self, param):
-        '''Cut a node, don't select it. Try to keep selection, then return the selected node that remains'''
-        w_ap = param["ap"]
-        if w_ap:
-            w_p = self._ap_to_p(w_ap)
-            if w_p:
-                if w_p == self.commander.p:
-                    self.commander.cutOutline()  # already on this node, so cut it
-                else:
-                    oldPosition = self.commander.p  # not same node, save position to possibly return to
-                    self.commander.selectPosition(w_p)
-                    self.commander.cutOutline()
-                    if self.commander.positionExists(oldPosition):
-                        # select if old position still valid
-                        self.commander.selectPosition(oldPosition)
-                    else:
-                        oldPosition._childIndex = oldPosition._childIndex-1
-                        # Try again with childIndex decremented
-                        if self.commander.positionExists(oldPosition):
-                            # additional try with lowered childIndex
-                            self.commander.selectPosition(oldPosition)
-                # in both cases, return selected node
-                return self._outputPNode(self.commander.p)
-            else:
-                # default empty
-                return self._outputError("Error in cutPNode no w_p node found")
-        else:
-            return self._outputError("Error in cutPNode no param node")
-
-    #@+node:ekr.20210611084045.118: *4* delete_node
-    def delete_node(self, param):
-        '''Delete a node, don't select it. Try to keep selection, then return the selected node that remains'''
-        w_ap = param["ap"]
-        if w_ap:
-            w_p = self._ap_to_p(w_ap)
-            if w_p:
-                if w_p == self.commander.p:
-                    self.commander.deleteOutline()  # already on this node, so delete it
-                else:
-                    oldPosition = self.commander.p  # not same node, save position to possibly return to
-                    self.commander.selectPosition(w_p)
-                    self.commander.deleteOutline()
-                    if self.commander.positionExists(oldPosition):
-                        # select if old position still valid
-                        self.commander.selectPosition(oldPosition)
-                    else:
-                        oldPosition._childIndex = oldPosition._childIndex-1
-                        # Try again with childIndex decremented
-                        if self.commander.positionExists(oldPosition):
-                            # additional try with lowered childIndex
-                            self.commander.selectPosition(oldPosition)
-                # in both cases, return selected node
-                return self._outputPNode(self.commander.p)
-            else:
-                # default empty
-                return self._outputError("Error in deletePNode no w_p node found")
-        else:
-            return self._outputError("Error in deletePNode no param node")
-
-    #@+node:ekr.20210611084045.119: *4* insert_node
-    def insert_node(self, param):
-        '''Insert a node at given node, then select it once created, and finally return it'''
-        w_ap = param["ap"]
-        if w_ap:
-            w_p = self._ap_to_p(w_ap)
-            if w_p:
-                w_bunch = self.commander.undoer.beforeInsertNode(w_p)
-                w_newNode = w_p.insertAfter()
-                w_newNode.setDirty()
-                self.commander.undoer.afterInsertNode(
-                    w_newNode, 'Insert Node', w_bunch)
-                self.commander.selectPosition(w_newNode)
-                # in both cases, return selected node
-                return self._outputPNode(self.commander.p)
-            else:
-                # default empty
-                return self._outputError("Error in insertPNode no w_p node found")
-        else:
-            return self._outputError("Error in insertPNode no param node")
-
-    #@+node:ekr.20210611084045.120: *4* insert_named_node
-    def insert_named_node(self, param):
-        '''Insert a node at given node, set its headline, select it and finally return it'''
-        w_newHeadline = param['name']
-        w_ap = param["ap"]
-        if w_ap:
-            w_p = self._ap_to_p(w_ap)
-            if w_p:
-                w_bunch = self.commander.undoer.beforeInsertNode(w_p)
-                w_newNode = w_p.insertAfter()
-                # set this node's new headline
-                w_newNode.h = w_newHeadline
-                w_newNode.setDirty()
-                self.commander.undoer.afterInsertNode(
-                    w_newNode, 'Insert Node', w_bunch)
-                self.commander.selectPosition(w_newNode)
-                # in any case, return selected node
-                return self._outputPNode(self.commander.p)
-            else:
-                # default empty
-                return self._outputError("Error in insertNamedPNode no w_p node found")
-        else:
-            return self._outputError("Error in insertNamedPNode no param node")
-
-    #@+node:ekr.20210611084045.121: *4* undo
-    def undo(self, param):
-        '''Undo last un-doable operation'''
-        if self.commander.undoer.canUndo():
-            self.commander.undoer.undo()
-        # return selected node when done
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.122: *4* redo
-    def redo(self, param):
-        '''Undo last un-doable operation'''
-        if self.commander.undoer.canRedo():
-            self.commander.undoer.redo()
-        # return selected node when done
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.123: *4* test
-    def test(self, param):
-        '''Utility test function for debugging'''
-        print("Called test")
-        return self.sendLeoBridgePackage({'testReturnedKey': 'testReturnedValue'})
-        # return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.124: *3* Outline and Body Interaction
-    #@+node:ekr.20210611084045.125: *4* page_up
-    def page_up(self, param):
-        """Selects a node a couple of steps up in the tree to simulate page up"""
-        n = param.get("n", 3)
-        for z in range(n):
-            self.commander.selectVisBack()
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.126: *4* page_down
-    def page_down(self, param):
-        """Selects a node a couple of steps down in the tree to simulate page down"""
-        n = param.get("n", 3)
-        for z in range(n):
-            self.commander.selectVisNext()
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.127: *4* get_body_states
-    def get_body_states(self, p_ap):
+    #@+node:ekr.20210209055518.1: *5* server.get_all_server_commands
+    def get_all_server_commands(self, package):
         """
-        Finds the language in effect at top of body for position p,
-        return type is lowercase 'language' non-empty string.
-        Also returns the saved cursor position from last time node was accessed.
-
-        The cursor positions are given as {"line": line, "col": col, "index": i}
-        with line and col along with a redundant index for convenience and flexibility.
+        Public server method:
+        Return the names of all callable public methods of the server.
         """
-        if not p_ap:
-            return self._outputError("Error in getLanguage, no param p_ap")
+        tag = 'get_all_server_commands'
+        names = self._get_all_server_commands()
+        if self.log_flag:  # pragma: no cover
+            print(f"\n{tag}: {len(names)} server commands\n")
+            g.printObj(names, tag=tag)
+        return self._make_response({"server-commands": names})
 
-        w_p = self._ap_to_p(p_ap)
-        if not w_p:
-            print(
-                "in GBS -> P NOT FOUND gnx:" + p_ap['gnx'] + " using self.commander.p gnx: " + self.commander.p.v.gnx)
-            w_p = self.commander.p
-
-        w_wrapper = self.commander.frame.body.wrapper
-
-        defaultPosition = {"line": 0, "col": 0, "index": 0}
-        states = {
-            'language': 'plain',
-            # See BodySelectionInfo interface in types.d.ts
-            'selection': {
-                "gnx": w_p.v.gnx,
-                "scroll": 0,
-                # "scroll": {
-                #     "start": defaultPosition,
-                #     "end": defaultPosition
-                # },
-                "insert": defaultPosition,
-                "start": defaultPosition,
-                "end": defaultPosition
-            }
-        }
-
-        if w_p:
-            c, g = self.commander, self.g
-            aList = g.get_directives_dict_list(w_p)
-            d = g.scanAtCommentAndAtLanguageDirectives(aList)
-
-            language = (
-                d and d.get('language') or
-                g.getLanguageFromAncestorAtFileNode(w_p) or
-                c.config.getString('target-language') or
-                'plain'
-            )
-
-            if w_p.v.scrollBarSpot is None:
-                w_scroll = 0
-            else:
-                w_scroll = w_p.v.scrollBarSpot
-
-            if w_p.v.insertSpot is None:
-                w_active = 0
-            else:
-                w_active = w_p.v.insertSpot
-            if w_p.v.selectionStart is None:
-                w_start = 0
-            else:
-                w_start = w_p.v.selectionStart
-
-            if w_p.v.selectionLength is None:
-                w_length = 0
-            else:
-                w_length = w_p.v.selectionLength
-
-            w_end = w_start + w_length
-
-            # get selection from wrapper instead if its the selected node
-            if self.commander.p.v.gnx == w_p.v.gnx:
-                # print("in GBS -> SAME AS self.commander.p SO USING FROM WRAPPER")
-                w_active = w_wrapper.getInsertPoint()
-                w_start, w_end = w_wrapper.getSelectionRange(True)
-                w_scroll = w_wrapper.getYScrollPosition()
-
-                w_activeI, w_activeRow, w_activeCol = c.frame.body.wrapper.toPythonIndexRowCol(
-                    w_active)
-                w_startI, w_startRow, w_startCol = c.frame.body.wrapper.toPythonIndexRowCol(
-                    w_start)
-                w_endI, w_endRow, w_endCol = c.frame.body.wrapper.toPythonIndexRowCol(
-                    w_end)
-            else:
-                print("NOT SAME AS self.commander.p SO USING FROM w_p.v")
-                w_activeI, w_startI, w_endI = w_active, w_start, w_end
-                w_activeRow, w_activeCol = g.convertPythonIndexToRowCol(
-                    w_p.v.b, w_active)
-                w_startRow, w_startCol = g.convertPythonIndexToRowCol(
-                    w_p.v.b, w_start)
-                w_endRow, w_endCol = g.convertPythonIndexToRowCol(
-                    w_p.v.b, w_end)
-
-            states = {
-                'language': language.lower(),
-                'selection': {
-                    "gnx": w_p.v.gnx,
-                    "scroll": w_scroll,  # w_scroll was kept as-is ?
-                    "insert": {"line": w_activeRow, "col": w_activeCol, "index": w_activeI},
-                    "start": {"line": w_startRow, "col": w_startCol, "index": w_startI},
-                    "end": {"line": w_endRow, "col": w_endCol, "index": w_endI}
-                }
-            }
-        return self.sendLeoBridgePackage(states)
-
-    #@+node:ekr.20210611084045.128: *4* get_children & helper
-    def get_children(self, p_ap):
-        '''EMIT OUT list of children of a node'''
-        if p_ap:
-            w_p = self._ap_to_p(p_ap)
-            if w_p and w_p.hasChildren():
-                return self._outputPNodes(w_p.children())
-            else:
-                return self._outputPNodes([])  # default empty array
-        else:
-            if self.commander.hoistStack:
-                return self._outputPNodes([self.commander.hoistStack[-1].p])
-            else:
-                # this outputs all Root Children
-                return self._outputPNodes(self._yieldAllRootChildren())
-
-    #@+node:ekr.20210611084045.129: *5* _yieldAllRootChildren
-    def _yieldAllRootChildren(self):
-        '''Return all root children P nodes'''
-        p = self.commander.rootPosition()
-        while p:
-            yield p
-            p.moveToNext()
-
-    #@+node:ekr.20210611084045.130: *4* get_parent
-    def get_parent(self, p_ap):
-        '''EMIT OUT the parent of a node, as an array, even if unique or empty'''
-        if p_ap:
-            w_p = self._ap_to_p(p_ap)
-            if w_p and w_p.hasParent():
-                return self._outputPNode(w_p.getParent())  # if not root
-        return self._outputPNode()  # default empty for root as default
-
-    #@+node:ekr.20210611084045.131: *4* get_all_gnx
-    def get_all_gnx(self, param):
-        '''Get gnx array from all unique nodes'''
-        w_all_gnx = [
-            p.v.gnx for p in self.commander.all_unique_positions(copy=False)]
-        return self.sendLeoBridgePackage({"gnx": w_all_gnx})
-
-    #@+node:ekr.20210611084045.132: *4* get_body
-    def get_body(self, p_gnx):
-        '''EMIT OUT body of a node'''
-        # TODO : if not found, send code to prevent unresolved promise if 'document switch' occurred shortly before
-        if p_gnx:
-            w_v = self.commander.fileCommands.gnxDict.get(p_gnx)  # vitalije
-            if w_v:
-                if w_v.b:
-                    return self._outputBodyData(w_v.b)
-                else:
-                    return self._outputBodyData()  # default "" empty string
-        # Send as empty to fix unresolved promise if 'document switch' occurred shortly before
-        return self._outputBodyData()
-
-    #@+node:ekr.20210611084045.133: *4* get_body_length
-    def get_body_length(self, p_gnx):
-        '''EMIT OUT body string length of a node'''
-        if p_gnx:
-            w_v = self.commander.fileCommands.gnxDict.get(p_gnx)  # vitalije
-            if w_v and w_v.b:
-                # Length in bytes, not just by character count.
-                return self.sendLeoBridgePackage({"len": len(w_v.b.encode('utf-8'))})
-        # TODO : May need to signal inexistent by self.sendLeoBridgePackage()
-        return self.sendLeoBridgePackage({"len": 0})  # empty as default
-
-    #@+node:ekr.20210611084045.134: *4* set_body
-    def set_body(self, param):
-        '''Change Body text of a v node'''
-        w_gnx = param['gnx']
-        w_body = param['body']
-        for w_p in self.commander.all_positions():
-            if w_p.v.gnx == w_gnx:
-                # TODO : Before setting undo and trying to set body, first check if different than existing body
-                w_bunch = self.commander.undoer.beforeChangeNodeContents(
-                    w_p)  # setup undoable operation
-                w_p.v.setBodyString(w_body)
-                self.commander.undoer.afterChangeNodeContents(
-                    w_p, "Body Text", w_bunch)
-                if self.commander.p.v.gnx == w_gnx:
-                    self.commander.frame.body.wrapper.setAllText(w_body)
-                if not self.commander.isChanged():
-                    self.commander.setChanged()
-                if not w_p.v.isDirty():
-                    w_p.setDirty()
-                break
-        # additional forced string setting
-        if w_gnx:
-            w_v = self.commander.fileCommands.gnxDict.get(w_gnx)  # vitalije
-            if w_v:
-                w_v.b = w_body
-        return self._outputPNode(self.commander.p)  # return selected node
-
-    #@+node:ekr.20210611084045.135: *4* get_focus
-    def get_focus(self, param):
+    def _get_all_server_commands(self):
         """
-        Return a representation of the focused widget,
-        one of ("body", "tree", "headline", repr(the_widget)).
+        Private server method:
+        Return the names of all callable public methods of the server.
         """
-        w = self.g.app.gui.get_focus()
-        focus = self.g.app.gui.widget_name(w)
-        return self.sendLeoBridgePackage({"focus": focus})
+        members = inspect.getmembers(self, inspect.ismethod)
+        return sorted([name for (name, value) in members if not name.startswith('_')])
+    #@+node:ekr.20210202110128.52: *5* server.init_connection
+    def _init_connection(self, web_socket):  # pragma: no cover (tested in client).
+        """Begin the connection."""
+        self.web_socket = web_socket
+        self.loop = asyncio.get_event_loop()
 
-    #@+node:ekr.20210611084045.136: *4* set_selection
-    def set_selection(self, param):
-        '''
-        Set cursor position and scroll position along with selection start and end.
-
-        Positions can be sent as {"col":int, "line" int} dict
-        or as numbers directly for convenience.
-
-        (For the currently selected node's body, if gnx matches only)
-        Save those values on the commander's body "wrapper"
-        See BodySelectionInfo interface in types.d.ts
-        '''
-        w_same = False  # Flag for actually setting values in the wrapper, if same gnx.
-        w_wrapper = self.commander.frame.body.wrapper
-        w_gnx = param['gnx']
-        w_body = ""
-        w_v = None
-        if self.commander.p.v.gnx == w_gnx:
-            # print('Set Selection! OK SAME GNX: ' + self.commander.p.v.gnx)
-            w_same = True
-            w_v = self.commander.p.v
-        else:
-            # ? When navigating rapidly - Check if this is a bug - how to improve
-            print('Set Selection! NOT SAME GNX: selected:' +
-                  self.commander.p.v.gnx + ', package:' + w_gnx)
-            w_v = self.commander.fileCommands.gnxDict.get(w_gnx)
-
-        if not w_v:
-            print('ERROR : Set Selection! NOT SAME Leo Document')
-            # ! FAILED (but return as normal)
-            return self._outputPNode(self.commander.p)
-
-        w_body = w_v.b
-        f_convert = self.g.convertRowColToPythonIndex
-        w_active = param['insert']
-        w_start = param['start']
-        w_end = param['end']
-
-        # no convertion necessary, its given back later
-        w_scroll = param['scroll']
-
-        # IF sent as number use as is - no conversion needed
-        if type(w_active) == int:
-            w_insert = w_active
-            w_startSel = w_start
-            w_endSel = w_end
-        else:
-            w_insert = f_convert(
-                w_body, w_active['line'], w_active['col'])
-            w_startSel = f_convert(
-                w_body, w_start['line'], w_start['col'])
-            w_endSel = f_convert(
-                w_body, w_end['line'], w_end['col'])
-
-        # print("setSelection (same as selected): " + str(w_same) + " w_insert " + str(w_insert) +
-        #       " w_startSel " + str(w_startSel) + " w_endSel " + str(w_endSel))
-
-        # If it's the currently selected node set the wrapper's states too
-
-        if w_same:
-            w_wrapper.setSelectionRange(w_startSel, w_endSel, w_insert)
-            w_wrapper.setYScrollPosition(w_scroll)
-        else:
-            pass
-
-        # Set for v node no matter what
-        w_v.scrollBarSpot = w_scroll
-        w_v.insertSpot = w_insert
-        w_v.selectionStart = w_startSel
-        w_v.selectionLength = (
-            w_endSel - w_startSel) if w_endSel > w_startSel else 0
-
-        # When switching nodes, Leo's core saves the insert point, selection,
-        # and vertical scroll position in the old (unselected) vnode. From v.init:
-
-        # self.insertSpot = None
-        #     # Location of previous insert point.
-        # self.scrollBarSpot = None
-        #     # Previous value of scrollbar position.
-        # self.selectionLength = 0
-        #     # The length of the selected body text.
-        # self.selectionStart = 0
-        #         # The start of the selected body text.
-
-        # output selected node as 'ok'
-        return self._outputPNode(self.commander.p)
-
-    #@+node:ekr.20210611084045.137: *4* set_headline
-    def set_headline(self, param):
-        '''Change Headline of a node'''
-        w_newHeadline = param['name']
-        w_ap = param["ap"]
-        if w_ap:
-            w_p = self._ap_to_p(w_ap)
-            if w_p:
-                # set this node's new headline
-                w_bunch = self.commander.undoer.beforeChangeNodeContents(w_p)
-                w_p.h = w_newHeadline
-                self.commander.undoer.afterChangeNodeContents(
-                    w_p, 'Change Headline', w_bunch)
-                return self._outputPNode(w_p)
-        return self._outputError("Error in setNewHeadline")
-
-    #@+node:ekr.20210611084045.138: *4* set_current_position & helper
-    def set_current_position(self, p_ap):
-        '''Select a node, or the first one found with its GNX'''
-        if p_ap:
-            w_p = self._ap_to_p(p_ap)
-            if w_p:
-                if self.commander.positionExists(w_p):
-                    # set this node as selection
-                    self.commander.selectPosition(w_p)
-                else:
-                    w_foundPNode = self._findPNodeFromGnx(p_ap['gnx'])
-                    if w_foundPNode:
-                        self.commander.selectPosition(w_foundPNode)
-                    else:
-                        print("Set Selection node does not exist! ap was:" +
-                              json.dumps(p_ap), flush=True)
-        # * return the finally selected node
-        if self.commander.p:
-            return self._outputPNode(self.commander.p)
-        else:
-            return self._outputPNode()
-
-    #@+node:ekr.20210611084045.139: *5* _findPNodeFromGnx
-    def _findPNodeFromGnx(self, p_gnx):
-        '''Return first p node with this gnx or false'''
-        for p in self.commander.all_unique_positions():
-            if p.v.gnx == p_gnx:
-                return p
-        return False
-
-    #@+node:ekr.20210611084045.140: *4* expand_node
-    def expand_node(self, p_ap):
-        '''Expand a node'''
-        if p_ap:
-            w_p = self._ap_to_p(p_ap)
-            if w_p:
-                w_p.expand()
-        return self.sendLeoBridgePackage()  # Just send empty as 'ok'
-
-    #@+node:ekr.20210611084045.141: *4* contract_node
-    def contract_node(self, p_ap):
-        '''Collapse a node'''
-        if p_ap:
-            w_p = self._ap_to_p(p_ap)
-            if w_p:
-                w_p.contract()
-        return self.sendLeoBridgePackage()  # Just send empty as 'ok'
-
-    #@+node:ekr.20210611084045.142: *3* leoFlexx Conversion Functions
-    #@+node:ekr.20210611084045.143: *4* _create_gnx_to_vnode
-    def _create_gnx_to_vnode(self):
-        '''Make the first gnx_to_vnode array with all unique nodes'''
-        t1 = time.process_time()
-        self.gnx_to_vnode = {
-            v.gnx: v for v in self.commander.all_unique_nodes()}
-        # This is likely the only data that ever will be needed.
-        if 0:
-            print('app.create_all_data: %5.3f sec. %s entries' % (
-                (time.process_time()-t1), len(list(self.gnx_to_vnode.keys()))), flush=True)
-        self._test_round_trip_positions()
-
-    #@+node:ekr.20210611084045.144: *4* _test_round_trip_positions
-    def _test_round_trip_positions(self):
-        '''(From Leo plugin leoflexx.py) Test the round tripping of p_to_ap and ap_to_p.'''
-        # Bug fix: p_to_ap updates app.gnx_to_vnode. Save and restore it.
-        old_d = self.gnx_to_vnode.copy()
-        old_len = len(list(self.gnx_to_vnode.keys()))
-        # t1 = time.process_time()
-        qtyAllPositions = 0
-        for p in self.commander.all_positions():
-            qtyAllPositions += 1
-            ap = self._p_to_ap(p)
-            p2 = self._ap_to_p(ap)
-            assert p == p2, (repr(p), repr(p2), repr(ap))
-        gnx_to_vnode = old_d
-        new_len = len(list(gnx_to_vnode.keys()))
-        assert old_len == new_len, (old_len, new_len)
-        # print('Leo file opened. Its outline contains ' + str(qtyAllPositions) + " nodes positions.", flush=True)
-        # print(('Testing app.test_round_trip_positions for all nodes: Total time: %5.3f sec.' % (time.process_time()-t1)), flush=True)
-
-    #@+node:ekr.20210611084045.145: *4* _ap_to_p
+    #@+node:ekr.20210205103759.1: *5* server.shut_down
+    def shut_down(self, package):
+        """Shut down the server."""
+        tag = 'shut_down'
+        n = len(g.app.commanders())
+        if n:  # pragma: no cover
+            raise ServerError(f"{tag}: {n} open outlines")
+        raise TerminateServer("client requested shut down")
+    #@+node:ekr.20210204154548.1: *3* server:server utils
+    #@+node:ekr.20210202110128.85: *4* server._ap_to_p
     def _ap_to_p(self, ap):
-        '''
-        (From Leo plugin leoflexx.py) Convert an archived position to a true Leo position.
-        Return false if no key
-        '''
-        childIndex = ap['childIndex']
+        """
+        Convert ap (archived position, a dict) to a valid Leo position.
+        Raise ServerError on any kind of error.
+        """
+        tag = '_ap_to_p'
+        c = self._check_c()
+        gnx_d = c.fileCommands.gnxDict
+        outer_stack = ap.get('stack')
+        if outer_stack is None:  # pragma: no cover.
+            raise ServerError(f"{tag}: no stack in ap: {ap!r}")
+        if not isinstance(outer_stack, (list, tuple)):  # pragma: no cover.
+            raise ServerError(f"{tag}: stack must be tuple or list: {outer_stack}")
+        
+        def d_to_childIndex_v (d):
+            """Helper: return childIndex and v from d ["childIndex"] and d["gnx"]."""
+            childIndex = d.get('childIndex')
+            if childIndex is None:  # pragma: no cover.
+                raise ServerError(f"{tag}: no childIndex in {d}")
+            try:
+                childIndex = int(childIndex)
+            except Exception:  # pragma: no cover.
+                raise ServerError(f"{tag}: bad childIndex: {childIndex!r}")
+            gnx = d.get('gnx')
+            if gnx is None:  # pragma: no cover.
+                raise ServerError(f"{tag}: no gnx in {d}.")
+            v = gnx_d.get(gnx)
+            if v is None:  # pragma: no cover.
+                raise ServerError(f"{tag}: gnx not found: {gnx!r}")
+            return childIndex, v
+        #
+        # Compute p.childIndex and p.v.
+        childIndex, v = d_to_childIndex_v(ap)
+        #
+        # Create p.stack.
+        stack = []
+        for stack_d in outer_stack:
+            stack_childIndex, stack_v = d_to_childIndex_v(stack_d)
+            stack.append((stack_v, stack_childIndex))
+        #
+        # Make p and check p.
+        p = Position(v, childIndex, stack)
+        if not c.positionExists(p):  # pragma: no cover.
+            print(
+                f"{tag}: Bad ap: {ap!r}\n"
+                # f"{tag}: position: {p!r}\n"
+                f"{tag}: v {v!r} childIndex: {childIndex!r}\n"
+                f"{tag}: stack: {stack!r}")
+            raise ServerError(f"{tag}: p does not exist in {c.shortFileName()}")
+        return p
+    #@+node:ekr.20210207054237.1: *4* server._check_c
+    def _check_c(self):
+        """Return self.c or raise ServerError if self.c is None."""
+        tag = '_check_c'
+        c = self.c
+        if not c:  # pragma: no cover
+            raise ServerError(f"{tag}: no open commander")
+        return c
+    #@+node:ekr.20210211131234.1: *4* server._check_outline
+    def _check_outline(self, c):
+        """Check self.c for consistency."""
+        # Check that all positions exist.
+        self._check_outline_positions(c)
+        # Test round-tripping.
+        self._test_round_trip_positions(c)
+    #@+node:ekr.20210211131827.1: *4* server._check_outline_positions
+    def _check_outline_positions(self, c):
+        """Verify that all positions in c exist."""
+        tag = '_check_outline_positions'
+        for p in c.all_positions(copy=False):
+            if not c.positionExists(p):  # pragma: no cover
+                message = f"{tag}: position {p!r} does not exist in {c.shortFileName()}"
+                print(message)
+                self._dump_position(p)
+                raise ServerError(message)
+    #@+node:ekr.20210209062536.1: *4* server._do_leo_command
+    def _do_leo_command(self, action, package):
+        """
+        Execute the leo command given by package ["leo-command-name"].
+        
+        The client must open an outline before calling this method.
+        """
+        # We *can* require self.c to exist, because:
+        # 1. all commands imply c.
+        # 2. The client must call open_file to set self.c.
+        tag = '_execute_leo_command'
+        c = self._check_c()
+        command_name = package.get("leo-command-name")
+        if not command_name:  # pragma: no cover
+            raise ServerError(f"{tag}: no 'leo-command-name' key in package")
+        if command_name in self.bad_commands_list:  # pragma: no cover
+            raise ServerError(f"{tag}: disallowed command: {command_name!r}")
+        func = c.commandsDict.get(command_name)
+        if not func:  # pragma: no cover
+            raise ServerError(f"{tag}: Leo command not found: {command_name!r}")
+        value = func(event={"c":c})
+        return self._make_response({"return-value": value})
+    #@+node:ekr.20210202110128.54: *4* server._do_message
+    def _do_message(self, d):
+        """
+        Handle d, a python dict representing the incoming request.
+        d must have at least the following keys:
+        
+        - "id": A positive integer.
+        - "action": A string, which is either:
+            - The name of public method of this class.
+            - "execute-leo-command".
+              d["package"]["leo-command-name"] should be the name of a Leo command.
+        
+        Return a dict, created by _make_response, containing least these keys:
 
-        try:
-            v = self.gnx_to_vnode[ap['gnx']]  # Trap this
-            stack = [
-                (self.gnx_to_vnode[d['gnx']], d['childIndex'])
-                for d in ap['stack']
-            ]
-        except Exception:
-            return False
+        - "id":         Same as the incoming id.
+        - "action":     Same as the incoming action.
+        - "commander":  A dict describing self.c.
+        - "node":       None, or an archived position describing self.c.p.
+        """
+        tag = '_do_message'
+        # Require "id" and "action" keys. The "package" key is optional.
+        id_ = d.get("id")
+        if id_ is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no id")
+        action = d.get("action")
+        if action is None:  # pragma: no cover
+            raise ServerError("f{tag}: no action")
+        package = d.get('package', {})
+        # Set log flag.
+        self.log_flag = package.get("log")
+        # Set the current_id and action ivars for _make_response.
+        self.current_id = id_
+        self.action = action
+        # Execute the requested action.
+        if action == "execute-leo-command":
+            func = self._do_leo_command
+        else:
+            func = self._do_server_command
+        result = func(action, package)
+        if result is None:  # pragma: no cover
+            raise ServerError(f"{tag}: no response: {action!r}")
+        return result
+    #@+node:ekr.20210209085438.1: *4* server._do_server_command
+    def _do_server_command(self, action, package):
+        tag = '_do_server_command'
+        # Disallow hidden methods.
+        if action.startswith('_'):  # pragma: no cover
+            raise ServerError(f"{tag}: action starts with '_': {action!r}")
+        # Find and execute the server method.
+        func = getattr(self, action, None)
+        if not func:
+            raise ServerError(f"{tag}: action not found: {action!r}")  # pragma: no cover
+        if not callable(func):
+            raise ServerError(f"{tag}: not callable: {func!r}")  # pragma: no cover
+        return func(package)
+    #@+node:ekr.20210211131707.1: *4* server._dump_*
+    def _dump_outline(self, c):  # pragma: no cover
+        """Dump the outline."""
+        tag = '_dump_outline'
+        print(f"{tag}: {c.shortFileName()}...\n")
+        for p in c.all_positions():
+            self._dump_position(p)
+        print('')
 
-        return leoNodes.position(v, childIndex, stack)
-
-    #@+node:ekr.20210611084045.146: *4* _p_to_ap
+    def _dump_position(self, p):  # pragma: no cover
+        level_s = ' ' * 2 * p.level()
+        print(f"{level_s}{p.childIndex():2} {p.v.gnx} {p.h}")
+    #@+node:ekr.20210202110128.51: *4* server._es & helper
+    def _es(self, s):  # pragma: no cover (tested in client).
+        """
+        Send a response that does not correspond to a request.
+        
+        The response *must* have an "async" key, but *not* an "id" key.
+        """
+        tag = '_es'
+        message = g.toUnicode(s)
+        package = {"async": "", "s": message}
+        response = json.dumps(package, separators=(',', ':'))
+        if self.loop:
+            self.loop.create_task(self._async_output(response))
+        else:
+            print(f"{tag}: Error loop not ready {message}")
+    #@+node:ekr.20210204145818.1: *5* server._async_output
+    async def _async_output(self, json):  # pragma: no cover (tested in server)
+        """Output json string to the web_socket"""
+        tag = '_async_output'
+        if self.web_socket:
+            await self.web_socket.send(bytes(json, 'utf-8'))
+        else:
+            g.trace(f"{tag}: no web socket. json: {json!r}")
+    #@+node:ekr.20210210081236.1: *4* server._get_p
+    def _get_p(self, package):
+        """Return _ap_to_p(package["ap"]) or c.p."""
+        tag = '_get_ap'
+        c = self.c
+        if not c:  # pragma: no cover
+            raise ServerError(f"{tag}: no c")
+        ap = package.get("ap")
+        if ap:
+            p = self._ap_to_p(ap)
+            if not p:  # pragma: no cover
+                raise ServerError(f"{tag}: no p")
+            if not c.positionExists(p):  # pragma: no cover
+                raise ServerError(f"{tag}: position does not exist. ap: {ap!r}")
+        if not c.p:  # pragma: no cover
+            raise ServerError(f"{tag}: no c.p")
+        return c.p
+    #@+node:ekr.20210211053733.1: *4* server._get_position_d
+    def _get_position_d(self, p):
+        """
+        Return a python dict containing:
+        - "node": self._p_to_ap(p).
+        - All *cheap* redraw data..
+        
+        Use get_ua to get p.ua *plus* all this redraw data.
+        
+        Note: v.computeIcon sets iconVal as follows:
+            v, val = self, 0
+            if v.hasBody(): val += 1
+            if v.isMarked(): val += 2
+            if v.isCloned(): val += 4
+            if v.isDirty(): val += 8
+        """
+        return {
+            "node": self._p_to_ap(p), # Contains p.gnx, p.childIndex and p.stack.
+            # The cheap redraw data...
+            "body-length": len(p.b),  # *Not* p.b.
+            "has-children": p.hasChildren(),  # *Not* p.children().
+            "has-ua": bool(p.v.u),  # *Not* p.v.u.
+            "headline": p.h,
+            "icon-val": p.v.iconVal,  # An int between 0 and 15.
+            "is-at-file": p.isAnyAtFileNode(),
+            "level": p.level(),  # Useful for debugging.
+        }
+    #@+node:ekr.20210206182638.1: *4* server._make_response
+    def _make_response(self, package=None):
+        """
+        Return a json string representing a response dict.
+        
+        The 'package' kwarg, if present, must be a python dict describing a
+        response. package may be an empty dict or None.
+        
+        The 'p' kwarg, if present, must be a position.
+        
+        First, this method creates a response (a python dict) containing all
+        the keys in the 'package' dict, with the following added keys:
+            
+        - "id":         The incoming id.
+        - "action":     The incoming action.
+        - "commander":  A dict describing self.c.
+        - "node":       None, or an archived position describing self.c.p.
+        
+        Finally, this method returns the json string corresponding to the
+        response.
+        """
+        tag = '_make_response'
+        c = self.c  # It is valid for c to be None.
+        if package is None:
+            package = {}
+        p = package.get("p")
+        if p:
+            del package ["p"]
+        # Raise an *internal* error if checks fail.
+        if isinstance(package, str):  # pragma: no cover
+            raise InternalServerError(f"{tag}: bad package kwarg: {package!r}")
+        if p and not isinstance(p, Position):  # pragma: no cover
+            raise InternalServerError(f"{tag}: bad p kwarg: {p!r}")
+        if p and not c:  # pragma: no cover
+            raise InternalServerError(f"{tag}: p but not c")
+        if p and not c.positionExists(p):  # pragma: no cover
+            raise InternalServerError(f"{tag}: p does not exist: {p!r}")
+        if c and not c.p:  # pragma: no cover
+            raise InternalServerError(f"{tag}: empty c.p")
+        #
+        # Always add these keys.
+        package ["id"] = self.current_id
+        package ["action"] = self.action
+        # The following keys are relevant only if there is an open commander.
+        if c:
+            # Allow commands, especially _get_redraw_d, to specify p!
+            p = p or c.p
+            package ["commander"] = {
+                "changed": c.isChanged(),
+                "file_name": c.fileName(), # Can be None for new files.
+            }
+            # Add all the node data, including:
+            # - "node": self._p_to_ap(p) # Contains p.gnx, p.childIndex and p.stack.
+            # - All the *cheap* redraw data for p.
+            redraw_d = self._get_position_d(p)
+            for key, value in redraw_d.items():
+                if key in package:  # pragma: no cover
+                    raise InternalServerError(f"{tag}: key {key!r} in package: {package!r}")
+                package [key] = value
+        if self.log_flag:  # pragma: no cover
+            g.printObj(package, tag=f"{tag} returns")
+        return json.dumps(package, separators=(',', ':')) 
+    #@+node:ekr.20210202110128.86: *4* server._p_to_ap
     def _p_to_ap(self, p):
-        '''(From Leo plugin leoflexx.py) Converts Leo position to a serializable archived position.'''
-        if not p.v:
-            print('app.p_to_ap: no p.v: %r %s' % (p), flush=True)
-            assert False
-        p_gnx = p.v.gnx
-        # * Expand gnx-vnode translation table for any new node encountered
-        if p_gnx not in self.gnx_to_vnode:
-            self.gnx_to_vnode[p_gnx] = p.v
-        # * necessary properties for outline
-        w_ap = {
+        """
+        Convert Leo position p to a serializable archived position.
+        
+        This returns only position-related data.
+        get_position_data returns all data needed to redraw the screen.
+        """
+        self._check_c()
+        stack = [{'gnx': v.gnx, 'childIndex': childIndex}
+            for (v, childIndex) in p.stack]
+        return {
             'childIndex': p._childIndex,
             'gnx': p.v.gnx,
-            'level': p.level(),
-            'headline': p.h,
-            'stack': [{
-                'gnx': stack_v.gnx,
-                'childIndex': stack_childIndex,
-                'headline': stack_v.h,
-            } for (stack_v, stack_childIndex) in p.stack],
+            'stack': stack,
         }
-        # TODO : Convert all those booleans into an 8 bit integer 'status' flag
-        # TODO : Send p.v.u as simple boolean flag and let user inspect ...
-        # TODO ... it with context menu command instead of hover tooltip.
-        if p.v.u:
-            w_ap['u'] = p.v.u
-        if bool(p.b):
-            w_ap['hasBody'] = True
-        if p.hasChildren():
-            w_ap['hasChildren'] = True
-        if p.isCloned():
-            w_ap['cloned'] = True
-        if p.isDirty():
-            w_ap['dirty'] = True
-        if p.isExpanded():
-            w_ap['expanded'] = True
-        if p.isMarked():
-            w_ap['marked'] = True
-        if p.isAnyAtFileNode():
-            w_ap['atFile'] = True
-        if p == self.commander.p:
-            w_ap['selected'] = True
-        return w_ap
-
-
+    #@+node:ekr.20210202110128.84: *4* serverver._test_round_trip_positions
+    def _test_round_trip_positions(self, c):  # pragma: no cover (tested in client).
+        """Test the round tripping of p_to_ap and ap_to_p."""
+        tag = '_test_round_trip_positions'
+        for p in c.all_unique_positions():
+            ap = self._p_to_ap(p)
+            p2 = self._ap_to_p(ap)
+            if p != p2:
+                self._dump_outline(c)
+                raise ServerError(f"{tag}: round-trip failed: ap: {ap!r}, p: {p!r}, p2: {p2!r}")
     #@-others
-#@+node:ekr.20210611084754.1: ** class TestLeoServer (unittest.TestCase)
+#@+node:ekr.20210208163018.1: ** class TestLeoServer (unittest.TestCase)
 class TestLeoServer (unittest.TestCase):  # pragma: no cover
     """Tests of LeoServer class."""
     request_number = 0
 
     #@+others
-    #@+node:ekr.20210611084754.2: *3* test: Setup and TearDown
+    #@+node:ekr.20210211085544.1: *3* test: Setup and TearDown
     @classmethod
     def setUpClass(cls):
         # Assume we are running in the leo-editor directory.
         # pylint: disable=import-self
-        from leo.core import leoserver
         global g, g_leoserver, g_server
         g_leoserver = leoserver
         g_server = leoserver.LeoServer(testing=True)
@@ -3315,7 +2247,7 @@ class TestLeoServer (unittest.TestCase):  # pragma: no cover
     def tearDown(self):
         g.unitTesting = False 
         
-    #@+node:ekr.20210611084754.3: *3* test._request
+    #@+node:ekr.20210208171819.1: *3* test._request
     def _request(self, action, package=None):
         server = self.server
         self.request_number += 1
@@ -3332,7 +2264,7 @@ class TestLeoServer (unittest.TestCase):  # pragma: no cover
         if log_flag:
             g.printObj(answer, tag=f"response to {action!r}")
         return answer
-    #@+node:ekr.20210611084754.4: *3* test.test_leo_commands
+    #@+node:ekr.20210210174801.1: *3* test.test_leo_commands
     def test_leo_commands (self):
         server = self.server
         table = [
@@ -3349,7 +2281,7 @@ class TestLeoServer (unittest.TestCase):  # pragma: no cover
                 self._request(action, package)
         finally:
             server.close_file({"filename": "xyzzy.leo"})
-    #@+node:ekr.20210611084754.5: *3* test.test_most_public_server_methods
+    #@+node:ekr.20210210102638.1: *3* test.test_most_public_server_methods
     def test_most_public_server_methods(self):
         server=self.server
         assert isinstance(server, g_leoserver.LeoServer), self.server
@@ -3408,7 +2340,7 @@ class TestLeoServer (unittest.TestCase):  # pragma: no cover
                             print(f"Exception in test_most_public_server_methods: {method_name!r} {e}")
         finally:
             server.close_file({"filename": test_dot_leo})
-    #@+node:ekr.20210611084754.6: *3* test.test_open_and_close
+    #@+node:ekr.20210208171319.1: *3* test.test_open_and_close
     def test_open_and_close(self):
         # server = self.server
         test_dot_leo = g.os_path_finalize_join(g.app.loadDir, '..', 'test', 'test.leo')
@@ -3437,7 +2369,7 @@ class TestLeoServer (unittest.TestCase):  # pragma: no cover
         ]
         for action, package in table:
             self._request(action, package)
-    #@+node:ekr.20210611084754.7: *3* test.test_find_commands
+    #@+node:ekr.20210212093613.1: *3* test.test_find_commands
     def test_find_commands(self):
         
         tag = 'test_find_commands'
@@ -3468,117 +2400,122 @@ class TestLeoServer (unittest.TestCase):  # pragma: no cover
             if log: g.printObj(answer, tag=f"{tag}:{method}: answer")
        
     #@-others
-#@+node:ekr.20210611084045.147: ** printAction
-def printAction(param):
-    '''Debugging tool that prints out called action if not in 'common-action' array'''
-    w_action = param["action"]
-    # print action if not getChild or getChildren or some other very common action
-    if w_action in commonActions:
-        pass
-    else:
-        print(f"*ACTION* {w_action}, id {param['id']}", flush=True)
-
-
-#@+node:ekr.20210611084045.148: ** main (server loop)
-def main():
-    '''python script for leo integration via leoBridge'''
+#@+node:ekr.20210202110128.88: ** function: main & helpers
+def main():  # pragma: no cover (tested in client)
+    """python script for leo integration via leoBridge"""
+    # from leo.core import leoGlobals as g
     global wsHost, wsPort
-    print("Starting Leobridgeserver.py (Launch with -h for help)", flush=True)
+    print("Starting LeoBridge... (Launch with -h for help)")
     # replace default host address and port if provided as arguments
+    #@+others
+    #@+node:ekr.20210202110128.90: *3* function: ws_handler (server)
+    async def ws_handler(websocket, path):
+        """
+        The web socket handler: server.ws_server.
 
-    args = None
-    try:
-        opts, args = getopt.getopt(sys.argv[1:], "ha:p:", [
-                                   "help", "address=", "port="])
-    except getopt.GetoptError:
-        print('leobridgeserver.py -a <address> -p <port>')
-        print('defaults to localhost on port 32125', flush=True)
-        if args:
-            print("unused args: " + str(args), flush=True)
-        sys.exit(2)
-    for opt, arg in opts:
-        if opt in ("-h", "--help"):
-            print('leobridgeserver.py -a <address> -p <port>')
-            print('defaults to localhost on port 32125', flush=True)
-            sys.exit()
-        elif opt in ("-a", "--address"):
-            wsHost = arg
-        elif opt in ("-p", "--port"):
-            wsPort = arg
-
-    # * start Server
-    integController = LeoBridgeIntegController()
-
-    # * This is a basic example loop
-    # async def asyncInterval(timeout):
-    #     dummyCounter = 0
-    #     strTimeout = str(timeout) + ' sec interval'
-    #     while True:
-    #         await asyncio.sleep(timeout)
-    #         dummyCounter = dummyCounter+1
-    #         await integController.asyncOutput("{\"interval\":" + str(timeout+dummyCounter) + "}")  # as number
-    #         print(strTimeout)
-
-    async def leoBridgeServer(websocket, path):
+        It must be a coroutine accepting two arguments: a WebSocketServerProtocol and the request URI.
+        """
+        tag = 'server'
+        trace = True
+        verbose = False
         try:
-            integController.initConnection(websocket)
-            # * Start by sending empty as 'ok'.
-            await websocket.send(integController.sendLeoBridgePackage())
-            integController.logSignon()
-            async for json_string_message in websocket:
-                messageObject = json.loads(json_string_message)
-                if messageObject and messageObject['action']:
-                    action = messageObject['action']
-                    param = messageObject['param']
-                    # printAction(w_param)  # Debug output
-                    # * Storing id of action in global var instead of passing as parameter.
-                    integController.setActionId(messageObject['id'])
-                    # ! functions called this way need to accept at least a parameter other than 'self'
-                    # ! See : getSelectedNode and getAllGnx
-                    if action[0] == "!":
-                        w_func = getattr(integController, action[1:], None)
-                        if not w_func:
-                            print(action)
-                        w_answer = w_func(param)
-                    elif action[0] == "_":
-                        integController._outputError(
-                            'Error : Command starting with underscore')
-                    else:
-                        # Attempt to execute the command directly on the commander/subcommander.
-                        w_answer = integController.leoCommand(
-                            action, param)
-                else:
-                    w_answer = "Error in processCommand"
-                    print(w_answer, flush=True)
-                await websocket.send(w_answer)
-        except websockets.exceptions.ConnectionClosedError:
-            print("Websocket connection closed", flush=True)
-        except Exception:
-            print('Exception in leobridgeserver.py!', flush=True)
-            # Like g.es_exception()...
-            typ, val, tb = sys.exc_info()
-            for line in traceback.format_exception(typ, val, tb):
-                print(line.rstrip(), flush=True)
-        finally:
-            asyncio.get_event_loop().stop()
-
-    localLoop = asyncio.get_event_loop()
-    start_server = websockets.serve(leoBridgeServer, wsHost, wsPort)
-    # localLoop.create_task(asyncInterval(5)) # Starts a test loop of async communication.
-    localLoop.run_until_complete(start_server)
-    # This SERVER_STARTED_TOKEN special string signals server startup success.
-    print(SERVER_STARTED_TOKEN + " at " + wsHost + " on port: " +
-          str(wsPort) + " [ctrl+c] to break", flush=True)
-    localLoop.run_forever()
-    print("Stopping leobridge server", flush=True)
-
-
+            controller._init_connection(websocket)
+            # Start by sending empty as 'ok'.
+            n = 0
+            async_n = 0
+            await websocket.send(controller._make_response())
+            # controller._sign_on()
+            async for json_message in websocket:
+                try:
+                    n += 1
+                    d = None
+                    trace = True  ## controller.trace
+                    d = json.loads(json_message)
+                    if trace and verbose:
+                        print(f"{tag}: got: {d}")
+                    elif trace:
+                        print(f"{tag}: got: {d.get('action')}")
+                    answer = controller._do_message(d)
+                except TerminateServer as e:
+                    raise websockets.exceptions.ConnectionClosed(code=1000, reason=e)
+                except ServerError as e:
+                    data = f"{d}" if d else f"json syntax error: {json_message!r}"
+                    error = f"{tag}:  ServerError: {e}...\n{tag}:  {data}"
+                    print("")
+                    print(error)
+                    print("")
+                    package = {
+                        "id": controller.current_id,
+                        "action": controller.action,
+                        "request": data,
+                        "ServerError": f"{e}",
+                    }  
+                    answer = json.dumps(package, separators=(',', ':'))
+                except InternalServerError as e:  # pragma: no cover
+                    print(f"{tag}: InternalServerError {e}")
+                    break
+                except Exception as e:  # pragma: no cover
+                    print(f"{tag}: Unexpected Exception! {e}")
+                    g.printObj(package, tag=f"message: {d}")
+                    g.print_exception()
+                    break
+                await websocket.send(answer)
+                if n in (3, 4, 7, 10):
+                    async_n += 1
+                    controller._es(f"async message {async_n}")
+        except websockets.exceptions.ConnectionClosedError as e:  # pragma: no cover
+            print(f"{tag}: closed error: {e}")
+        except websockets.exceptions.ConnectionClosed as e:
+            print(f"{tag}: closed normally: {e}")
+        # Don't call EventLoop.stop(). It terminates abnormally.
+            # asyncio.get_event_loop().stop()
+    #@+node:ekr.20210202110128.91: *3* function: get_args
+    def get_args():  # pragma: no cover
+        global wsHost, wsPort
+        args = None
+        try:
+            opts, args = getopt.getopt(sys.argv[1:], "help:", ["help", "address=", "port="])
+        except getopt.GetoptError:
+            print('leoserver.py -a <address> -p <port>')
+            print('defaults to localhost on port 32125')
+            if args:
+                print("unused args: " + str(args))
+            sys.exit(2)
+        for opt, arg in opts:
+            if opt in ("-h", "--help"):
+                print('leoserver.py -a <address> -p <port>')
+                print('defaults to localhost on port 32125')
+                sys.exit()
+            elif opt in ("-a", "--address"):
+                wsHost = arg
+            elif opt in ("-p", "--port"):
+                wsPort = arg
+        # Leave other options for unittest.
+        for opt, junk in opts:  # opts is a 2-tuple.
+            if opt in sys.argv:
+                sys.argv.remove(opt)
+        return wsHost, wsPort
+    #@-others
+    if '--unittest' in sys.argv:
+        sys.argv.remove('--unittest')
+        unittest.main()
+        return  # Make *sure* we don't start the server.
+    wsHost, wsPort = get_args()
+    signon = f"LeoBridge started at {wsHost} on port: {wsPort}. Ctrl+c to break"
+    print(signon)
+    # Open leoBridge.
+    controller = LeoServer()
+    # Start the server.
+    loop = asyncio.get_event_loop()  
+    server = websockets.serve(ws_handler=ws_handler, host=wsHost, port=wsPort)
+    loop.run_until_complete(server)
+    loop.run_forever()
 #@-others
 if __name__ == '__main__':
-    # Startup
+    # pytest will *not* execute this code.
     try:
         main()
     except KeyboardInterrupt:
-        print("\nKeyboard Interupt: Stopping leobridge server", flush=True)
+        print("\nKeyboard Interupt: Stopping leoserver.py")
         sys.exit()
 #@-leo
